% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.peak.error.R
\name{annual.peak.error}
\alias{annual.peak.error}
\title{Annual Peak Errors}
\usage{
annual.peak.error(sim, obs, rplot = T, add.line = T, add.labels = T)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{rplot}{boolean whether to generate plot (default TRUE)}

\item{add.line}{optionally adds a 1:1 line to the plot for reference
(default TRUE)}

\item{add.labels}{optionally adds labels for overpredict/underpredict on
right side axis (default TRUE)}
}
\value{
\item{df.peak.error}{data frame of the calculated peak errors}
}
\description{
annual.peak.error creates a plot of the annual observed and simulated peak
percent errors.
}
\details{
This function creates a plot of the percent errors in simulated peaks for
each water year. The peaks are calculated as the magnitude of the largest
event in each water year. Note that the annual.peak.error function is first
used to obtain the peaks in each year, then the percent errors are
calculated.

The percent errors are calculated as (QPsim-QPobs)/QPobs*100, where QP is
the peak flow event.

The sim and obs should be of time series (xts) format and are assumed to be
of the same length and time period. The flow series are assumed to be daily
flows with units of m3/s.

The R2 diagnostic is calculated for a fit with no intercept (in a perfect
fit the points are identical, and intercept is automatically zero).

The add.labels will add the labels of 'overprediction' and 'underprediction'
to the right hand side axis if set to TRUE. This is useful in interpreting
the plots.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(hydrograph.data)
sim <- hydrograph.data$hyd$Sub36
obs <- hydrograph.data$hyd$Sub36_obs

# create a plot of peak annual errors with default options
annual.peak.error(sim,obs,rplot=T,add.line=T)

# do not plot, just store the calculated peak errors
peak.errors <- annual.peak.error(sim,obs,rplot=F)
peak.errors

}
\seealso{
\code{\link{annual.peak.event}} to consider the timing of peak
events \code{\link{annual.peak.event.error}} to calculate errors in peak
events

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{annual}
\keyword{diagnostics}
\keyword{error}
\keyword{peak}
