% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.peak.event.R
\name{annual.peak.event}
\alias{annual.peak.event}
\title{Annual Peak Event Comparison}
\usage{
annual.peak.event(sim, obs, rplot = T, add.line = T, add.r2 = F,
  axis.zero = F)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{rplot}{boolean whether to generate plot (default TRUE)}

\item{add.line}{optionally adds a 1:1 line to the plot for reference
(default TRUE)}

\item{add.r2}{optionally computes the R2 and adds to plot (default FALSE)}

\item{axis.zero}{optionally sets the minimum volume on axes to zero (default
FALSE)}
}
\value{
\item{df.peak.event}{data frame of the calculated peak events}
}
\description{
annual.peak.event creates a plot of the annual observed and simulated peaks.
}
\details{
This function creates a scatterplot of the annual observed and simulated
peaks, calculated for each available water year of data (Oct 1st hardcoded)
within the two series provided; note that the difference between this and
the annual.peak function is that here the peak event simulated for the same
day as the peak event in observed data is used, instead of the largest
recorded simulated event. In some sense this captures the timing of the
event, i.e. the peak event must be simulated on the same day as the observed
peak to be captured well.

The sim and obs should be of time series (xts) format and are assumed to be
of the same length and time period. The flow series are assumed to be daily
flows with units of m3/s.

The R2 diagnostic is calculated for a fit with no intercept (in a perfect
fit the points are identical, and intercept is automatically zero).

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(hydrograph.data)
sim <- hydrograph.data$hyd$Sub36
obs <- hydrograph.data$hyd$Sub36_obs

# create a plot of annual peak events with default options
annual.peak.event(sim,obs,rplot=T,add.line=T,add.r2=F,axis.zero=F)

# plot with r2 and axes to zero; store results
peak.df <- annual.peak.event(sim,obs,rplot=T,add.line=T,add.r2=T,axis.zero=T)
peak.df

# store results without plotting
peak.df <- annual.peak.event(sim,obs,rplot=F)
peak.df

}
\seealso{
\code{\link{annual.peak}} to create a scatterplot of annual peaks
(consider the magnitude of peaks only)

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{annual}
\keyword{diagnostics}
\keyword{event}
\keyword{peak}
