% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.peak.timing.error.R
\name{annual.peak.timing.error}
\alias{annual.peak.timing.error}
\title{Annual Peak Timing Errors}
\usage{
annual.peak.timing.error(sim, obs, rplot = T, add.line = T,
  add.labels = T)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{rplot}{boolean whether to generate plot (default TRUE)}

\item{add.line}{optionally adds a 1:1 line to the plot for reference
(default TRUE)}

\item{add.labels}{optionally adds labels for early peak/late peaks on right
side axis (default TRUE)}
}
\value{
\item{df.peak.timing.error}{data frame of the calculated peak timing
errors}
}
\description{
annual.peak.timing.error creates a plot of the annual observed and simulated
peak timing errors.
}
\details{
This function creates a plot of the peak timing errors in simulated peaks
for each water year. The difference in days between the simulated peak and
observed peak are plotted (and/or returned in the data frame) for the water
year. This diagnostic is useful in determining how accurate the timing of
peak predictions is. Note that a large error in the number of days between
simulated and observed peaks indicates that the model predicted a larger
event at a different time of year, i.e. overestimated a different event or
underestimated the actual peak event, relative to the observed flow series.

The sim and obs should be of time series (xts) format and are assumed to be
of the same length and time period. The flow series are assumed to be daily
flows with units of m3/s. Note that a plot title is purposely omitted in
order to allow the automatic generation of plot titles.

The add.labels will add the labels of 'early peak' and 'late peak' to the
right hand side axis if set to TRUE. This is useful in interpreting the
plots. Note that values in this metric of less than zero indicate an early
prediction of the peak, and positive values mean a late prediction of the
peak (since the values are calculated as day index of simulated peak - day
index of observed peak).
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(hydrograph.data)
sim <- hydrograph.data$hyd$Sub36
obs <- hydrograph.data$hyd$Sub36_o

# create a plot of peak timing errors with defaults
annual.peak.timing.error(sim,obs)

# create the plot without labels or 1:1 line or labels
annual.peak.timing.error(sim,obs,add.line=F,add.labels=F)

# avoid plotting and
timing.errs <- annual.peak.timing.error(sim,obs,rplot=F)

}
\seealso{
\code{\link{annual.peak.event}} to consider the timing of peak
events \code{\link{annual.peak.event.error}} to calculate errors in peak
events

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{annual}
\keyword{diagnostics}
\keyword{error}
\keyword{peak}
\keyword{timing}
