% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.volume.R
\name{annual.volume}
\alias{annual.volume}
\title{Annual Volume Comparison}
\usage{
annual.volume(sim, obs, rplot = T, add.line = T, add.r2 = F,
  axis.zero = F)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{rplot}{boolean whether to generate plot (default TRUE)}

\item{add.line}{optionally adds a 1:1 line to the plot for reference
(default TRUE)}

\item{add.r2}{optionally computes the R2 and adds to plot (default FALSE)}

\item{axis.zero}{optionally sets the minimum volume on axes to zero (default
FALSE)}
}
\value{
\item{df.volume}{data frame of the calculated annual volumes}
}
\description{
annual.volume creates a plot of the annual observed and simulated volumes.
}
\details{
This function creates a scatterplot of the annual observed and simulated
volumes, calculated for each available water year of data (Oct 1st
hardcoded) within the two series provided. The sim and obs should be of time
series (xts) format and are assumed to be of the same length and time
period. Note that missing values in the observed series will impact the
volume estimation, and it is recommended that the NA values are filled in
prior to use of this function.

The R2 diagnostic is calculated for a fit with no intercept (in a perfect
fit the points are identical, and intercept is automatically zero).

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(hydrograph.data)
sim <- hydrograph.data$hyd$Sub36
obs <- hydrograph.data$hyd$Sub36_o

# create a plot of the annual volumes with defaults
annual.volume(sim,obs)

# create a plot of the annual volumes with r2 and axis set to zero
annual.volume(sim,obs,add.r2=T,axis.zero=T)

# store results of annual volumes
volumes <- annual.volume(sim,obs,rplot=F)

}
\seealso{
\code{\link{flow.scatterplot}} to create a scatterplot of flow
values

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{annual}
\keyword{diagnostics}
\keyword{volume}
