% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.runoff.coeff.R
\name{calc.runoff.coeff}
\alias{calc.runoff.coeff}
\title{Generate runoff coefficients upstream of gauges}
\usage{
calc.runoff.coeff(rvhfile, outdir,
  custfile = "PRECIP_Daily_Average_BySubbasin.csv",
  hydfile = "Hydrographs.csv", correct = FALSE)
}
\arguments{
\item{rvhfile}{rvh filename (absolute or relative to working directory, with .rvh extension)}

\item{outdir}{output directory containing PRECIP_Daily_Average_BySubbasin.csv (absolute or relative to working directory, with ending slash)}

\item{custfile}{(optional) file name (no directory) for Raven-generated custom output precip-by-subbasin file}

\item{hydfile}{(optional) file name (no directory) for Raven-generated hydrographs file}

\item{correct}{(optional) if TRUE, tries to correct runoff coefficient for missing data (assumes missing~0 flow)}
}
\value{
data frame with runoff coefficients of gauged basins
}
\description{
Reads model.rvh file and daily avg subbasin precip file (usually PRECIP_Daily_Average_BySubbasin.csv)
and generates data frame describing runoff coefficients of gauged basins and observation data coverage.
Uses precipitation from entire model run history.
Only determines runoff coefficient from available data - prone to overestimation with poor observation coverage.
}
\details{
requires igraph library
}
\examples{

  out<-calc.runoff.coeff("example.rvh","C:/TEMP/ouput/")

}
\seealso{
See also the \href{http://raven.uwaterloo.ca/}{Raven web site}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Raven}
\keyword{coefficient}
\keyword{runoff}
\keyword{rvh}
