% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_error.R
\name{capture_error}
\alias{capture_error}
\title{Custom stop function to intercept errors}
\usage{
capture_error()
}
\description{
Custom stop function to intercept errors
}
\examples{
options(error = ravenr::capture_error)
log('a')
}
