% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_exception.R
\name{capture_exception}
\alias{capture_exception}
\title{Capture Exception}
\usage{
capture_exception(exception, extra, level, tags,
  include_session_info)
}
\arguments{
\item{object}{A Sentry client}

\item{exception}{exception to catch}

\item{extra}{set extra context}

\item{level}{set level, warning or error for example}

\item{tags}{named list of tags}

\item{include_session_info}{whether to send platform and package list, takes up to 1s or more}
}
\description{
\code{capture_exception} Captures an exception and sends it to Sentry
}
