% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_warning.R
\name{capture_warning}
\alias{capture_warning}
\title{Custom stop function to intercept warnings}
\usage{
capture_warning(res)
}
\description{
Custom stop function to intercept warnings
}
\examples{
options(warning.expression = quote({ravenr::capture_warning(restart)}))
log(-1)
}
