% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{cmax}
\alias{cmax}
\title{cmax

cmax standards for column maximum. It applies the base::max function over
columns, which is advantageous for calculating the max within a column
rather than the max of the whole data frame.}
\usage{
cmax(x, na.rm = FALSE)
}
\arguments{
\item{x}{object to apply the max function to}

\item{na.rm}{whether to remove na values from the calculation}
}
\value{
x with the max value in each column determined
}
\description{
cmax

cmax standards for column maximum. It applies the base::max function over
columns, which is advantageous for calculating the max within a column
rather than the max of the whole data frame.
}
\examples{
data(rvn_hydrograph_data)
rvn_which_max_xts(rvn_hydrograph_data$hyd$Sub43_obs)

}
\seealso{
\code{\link{rvn_apply_wyearly}} where this function can be applied for the water year

The original [SO post](https://stackoverflow.com/questions/51623745/how-to-use-apply-daily-period-apply-for-calculating-maximum-per-column-in-xts-ti)

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{max}
\keyword{which}
\keyword{xts}
