% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.transparent.R
\name{col.transparent}
\alias{col.transparent}
\title{Add Transparency to Colours}
\usage{
col.transparent(colour, trans)
}
\arguments{
\item{colour}{time series containing columns you wish to reseasonalize. xts
object}

\item{trans}{integer describing the degree of transparency, from ~200
(slightly transparent) to <10 (very transparent)}
}
\value{
\item{res}{returned updated colour code with transparency}
}
\description{
col.transparent is used to adjust colour codes to introduce transparency
}
\examples{

# plot randomly distributed data
plot(rnorm(20),col='black')

# create a transparent blue colour for plotting
mycol <- col.transparent('blue',100)

# plot more random points in transparent blue colour
points(rnorm(20),col=mycol)

}
\seealso{
See original code on post in Stack Overflow
\href{http://stackoverflow.com/questions/12995683/any-way-to-make-plot-points-in-scatterplot-more-transparent-in-rmaking}{
plot points transparent in R}

\code{\link{iscolour}} for checking validity of colour codes
}
\keyword{colour}
\keyword{transparency}
