% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry_client.R
\name{create_sentry_client}
\alias{create_sentry_client}
\title{Create Sentry Client}
\usage{
create_sentry_client(dsn, user = NULL, version = NULL,
  log_errors = TRUE, log_warnings = TRUE)
}
\arguments{
\item{dsn}{set Data Source Name}

\item{user}{set user context}

\item{version}{protocol version}

\item{log_errors}{whether to send errors to sentry}

\item{log_warnings}{whether to send warnings to sentry}
}
\description{
\code{create_sentry_client} Creates a sentry client
}
