% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.rvt.R
\name{ECflow.rvt}
\alias{ECflow.rvt}
\title{EC Streamgauge File Conversion}
\usage{
ECflow.rvt(ff, subIDs, prd = NULL, stnNames = NULL,
  write.redirect = F, flip.number = F)
}
\arguments{
\item{ff}{WSC flow file in csv format}

\item{subIDs}{vector of subbasin IDs to correspond to the stations in ff}

\item{prd}{(optional) data period to use in .rvt file}

\item{stnNames}{(optional) character vector of alternative station names to
use}

\item{write.redirect}{(optional) write the :RedirectToFile commands in a
separate .rvt file}

\item{flip.number}{(optional) put the subID first in the .rvt filename}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
ECflow.rvt converts Environment Canada historical streamgauge data,
downloaded from the Water Survey Canada, into .rvt format files usable in
Raven.
}
\details{
This function takes a single WSC flow file and converts the flow data for
each station in the file into .rvt formatted files for a Raven model. If
multiple stations exist in the .csv file,

The file should be downloaded in a .csv Date-Data format with missing days
included. The download website is linked below. Any level data will be
ignored, although can be included in the file for stations with flow and
level data. If no data is found for a given station, an error will be
reported.

subIDs is required and should correspond to the subID to be used in the .rvt
file for each station in the ff file, in the order in which it will be read
in. If the subbasin is currently unknown, please supply a placeholder value.

prd is used by the xts formatted-data to restrict the data reported in .rvt
files, for each station, to this period. The prd should be defined in
"YYYY-MM-DD/YYYY-MM-DD" string format. If the period supplied results in an
empty time series (i.e. non-overlapping time periods), an error will be
thrown.

stnNames is an optional character vector to replace the EC station codes
found in the .csv file. If supplied, the vector must be of the same length
as the number of stations supplied and the subIDs vector. If not supplied,
the EC station codes will be used. Note that this does not impact model
function, only filename readability and station recognition.

write.redirect will print out the :RedirectToFile commands in a separate
file called, "flow_stn_redirect_text.rvt". These commands can be copied into
the main model's .rvt file to redirect to the produced time series files.

flip.number is a useful option to place the subID first in the filename.
This is often cleaner for organizing files in a folder, since the
alphabeticized order is not dependent on the station name, and the observed
files will be in one set.
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{
ff <- 'Daily__May-12-2017_02_00_53PM.csv'
ECflow.rvt(ff,subIDs=c(3,11))

# add custom station names, put subID number first in file
ECflow.rvt(ff,subIDs=c(3,11),stnNames<-c('Rob_Hill','Bob_River'),flip.number=T)
}

}
\seealso{
\code{\link{annual.peak.event}} to consider the timing of peak
events

Download EC streamgauge data from
\href{https://wateroffice.ec.gc.ca/search/historical_e.html}{WSC Historical
Data}
See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{conversion}
\keyword{flow}
\keyword{rvt}
\keyword{streamgauge}
