% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflowTidy.rvt.R
\name{ECflowTidy.rvt}
\alias{ECflowTidy.rvt}
\title{EC Streamgauge File Conversion from tidyhydat}
\usage{
ECflowTidy.rvt(indata, subIDs, prd = NULL, stnNames = NULL,
  write.redirect = F, flip.number = F)
}
\arguments{
\item{indata}{tibble of WSC flow data from tidyhydat's hy_daily_flows() 
function}

\item{subIDs}{vector of subbasin IDs to correspond to the stations in ff}

\item{prd}{(optional) data period to use in .rvt file}

\item{stnNames}{(optional) character vector of alternative station names to
use}

\item{write.redirect}{(optional) write the :RedirectToFile commands in a
separate .rvt file}

\item{flip.number}{(optional) put the subID first in the .rvt filename}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}

See also the \href{http://www.raven.uwaterloo.ca/}{Raven page}
}
\description{
ECflowTidy.rvt converts Environment Canada historical streamgauge data,
accessed via the tidyhydat package, into .rvt format files usable in
Raven.
}
\details{
This function takes a single flow tibble generated from tidyhydat and converts the flow data for
each station in the file into .rvt formatted files for a Raven model. If
multiple stations exist in the .csv file, multiple observation files are created

subIDs is required and should correspond to the subID to be used in the .rvt
file for each station in the ff file, in the order in which it will be read
in.

prd is used by the xts formatted-data to restrict the data reported in .rvt
files, for each station, to this period. The prd should be defined in
"YYYY-MM-DD/YYYY-MM-DD" string format. If the period supplied results in an
empty time series (i.e. non-overlapping time periods), an error will be
thrown.

stnNames is an optional character vector to replace the EC station codes
found in the HYDAT database. If supplied, the vector must be of the same length
as the number of stations supplied and the subIDs vector. If not supplied,
the EC station codes will be used. Note that this does not impact model
function, only filename readability and station recognition.

write.redirect will print out the :RedirectToFile commands in a separate
file called, "flow_stn_redirect_text.rvt". These commands can be copied into
the main model's .rvt file to redirect to the produced time series files.

flip.number is a useful option to place the subID first in the filename.
This is often cleaner for organizing files in a folder, since the
alphabeticized order is not dependent on the station name, and the observed
files will be in one set.
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{
data<-hy_daily_flows(station_number = c("05CB004","05CA002"),start_date = "1996-01-01", end_date = "2000-01-01")

stationnames<-c('Raven River','Lower James River')

ECflowTidy.rvt(data,subIDs=c(3,11),stnNames=stationnames,flip.number=T)
}

}
\keyword{Raven}
\keyword{conversion}
\keyword{flow}
\keyword{rvt}
\keyword{streamgauge}
\keyword{tidyhydat}
