% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhaustivemb.read.R
\name{exhaustivemb.read}
\alias{exhaustivemb.read}
\title{Read in Raven Exhaustive Mass Balance file}
\usage{
exhaustivemb.read(ff = NA, join.categories = T)
}
\arguments{
\item{ff}{full file path to the ExhaustiveMassBalance.csv file}

\item{join.categories}{boolean whether add to the category tag as a column
name prefix in exhaustivemb output (default TRUE)}
}
\value{
\item{exhaustivemb}{data frame from the file with standardized
names} \item{units}{vector corresponding to units of each column}
\item{categories}{vector corresponding to the storage category of each column}
}
\description{
watershed.read is used to read in the ExhaustiveMassBalance.csv file produced by
the modelling Framework Raven.
}
\details{
This function expects a full file path to the ExhaustiveMassBalance.csv file,
then reads in the file using read.csv. The main advantage of this functon is
renaming the columns to nicer names and extracting the units into something
much easier to read.

ff is the full file path of the ExhaustiveMassBalance.csv file. If the file is
located in the current working directory, then simply the name of the file
is sufficient.
}
\examples{
read in file from package (RavenR/inst/extdata)
ff <- system.file("extdata","ExhaustiveMassBalance_1.csv", package="RavenR")

# read in exhaustive MB file, create plot
exhaustiveMB.data <- RavenR::exhaustivemb.read(ff)
head(exhaustiveMB.data$exhaustivemb)
plot(exhaustiveMB.data$exhaustivemb$SURFACE_WATER.Infiltration,
ylab="Cumulative Surface Water Infiltration"
}
\seealso{
\code{\link{hyd.read}} for reading in the Hydrographs.csv file
\code{\link{watershedmeb.read}} for reading in the
WatershedMassEnergyBalance.csv file

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{balance}
\keyword{exhaustive}
\keyword{mass}
\keyword{read.csv}
