% Documentation file for external data, not generated by roxygen

\name{Nith_era5_sample}
\alias{Nith_era5_sample}
\title{ERA5 Sample Dataset for Nith Watershed}
\description{
ERA5 dataset in netcdf format, obtained as a sample dataset from the Global Water Futures (GWF) \href{https://tuna.cs.uwaterloo.ca/}{Cuizinart portal}.
}
\details{
Additional metadata on the ERA5 dataset may be found on the \href{https://gwf-uwaterloo.github.io/wamedex/#13079_iso.xml}{GWF metadata page}.
}
\seealso{
\code{\link{rvn_netcdf_to_gridshp}} to generate a grid shapefile from a netcdf file
}
\examples{
# access the file from the RavenR installation
system.file("extdata/Nith_era5_sample.nc", package="RavenR")
}
