% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.scatterplot.R
\name{flow.scatterplot}
\alias{flow.scatterplot}
\title{Scatterplot of model flows}
\usage{
flow.scatterplot(sim, obs, add.line = T, add.r2 = F, axis.zero = T)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{add.line}{optionally adds a 1:1 line to the plot for reference
(default TRUE)}

\item{add.r2}{optionally computes the R2 and adds to plot (default FALSE)}

\item{axis.zero}{optionally sets the minimum flow on axes to zero (default
TRUE)}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
flow.scatterplot creates a scatterplot of the simulated and observed flows.
Useful in diagnotic analysis of model outputs.
}
\details{
This function creates a scatterplot of flows.

The sim and obs should be of time series (xts) format. The flow series are
assumed to be daily flows with units of m3/s.

The R2 diagnostic is calculated for a fit with no intercept (in a perfect
fit the points are identical, and intercept is automatically zero). The R2
is calculated with the NA values removed

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(hydrograph.data)
sim <- hydrograph.data$hyd$Sub36
obs <- hydrograph.data$hyd$Sub36_ob

# plot the flow scatterplot, produce an R2 metric
flow.scatterplot(sim,obs,add.r2=T)

}
\seealso{
\code{\link{forcings.read}} for reading in the ForcingFunctions
file

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{diagnostics}
\keyword{flow}
\keyword{scatterplot}
