% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.spaghetti.R
\name{flow.spaghetti}
\alias{flow.spaghetti}
\title{Flow Spaghetti Plot}
\usage{
flow.spaghetti(flow, include.legend = T, colour.wheel = T)
}
\arguments{
\item{flow}{time series object of simulated flows}

\item{include.legend}{option to include a legend (default TRUE)}

\item{colour.wheel}{option to use a different colour for each year plotted
(default TRUE)}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
flow.spaghetti creates a spaghetti plot of the flow series provided.
}
\details{
This function creates a spaghetti plot of the annual flow series in each
year of data provided. The flows are plotted for each water year of data
available, set as October 1st.

Note that the plotting to the day of year is approximate in order to
simplify the plotting of leap years and non-leap years. The years are
plotted including day 366 and starting on day 274, regardless of whether it
is a leap year or not. This is likely without consequence in seeing the
trends between water years, however the user is warned of this deficiency.

The include.legend argument will add a legend to the right-hand side of the
plot, if true. The legend will be labelled as the year for each flow series,
where the year is the date-ending year (i.e. for the 2003-10-01 to
2004-09-30 water year, the flow series will be labelled as 2004). This
argument will also cause the plot margins to be adjusted so that the legend
can be included. Note that the current setup to accomodate approximately 15
years of flow series in the legend, beyond that the legend may not be
aesthetically pleasing.

The colour.wheel argument, if TRUE, will plot each water year in a different
randomly sampled colour. If FALSE, all years will be plotted in black. If
colour.wheel is FALSE, the legend is recommended to be set to FALSE as well.

The flow series provided should be of time series (xts) format.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(hydrograph.data)
sim <- hydrograph.data$hyd$Sub36
sim2 <- hydrograph.data$hyd$Sub43

# create spaghetti plot of simulated flows
flow.spaghetti(sim)

# create spaghetti plot without legend or different colours
flow.spaghetti(sim2,include.legend=F,colour.wheel=F)

}
\seealso{
\code{\link{flow.scatterplot}} to create a scatterplot of flow
values

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{diagnostics}
\keyword{flow}
\keyword{spaghetti}
