% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowdurcurve.plot.R
\name{flowdurcurve.plot}
\alias{flowdurcurve.plot}
\title{Plots summary of watershed forcing functions}
\usage{
flowdurcurve.plot(hydrograph, prd = NULL, seasonal = "F")
}
\arguments{
\item{hydrograph}{hydrograph object from hyd.extract function}

\item{prd}{(optional) time period over which the plot is generated}

\item{seasonal}{(optional) add the winter and summer FDC}
}
\description{
flowdurcurve.plot generation a flow duration curve plot.
}
\details{
This function creates a flow duration curve using the hyd.extract obejct for
a given basin. The hydrograph object passed should be the output from the
hyd.extract function, which has attributes for sim and obs; if the obs is
NULL, only the sim FDC will be plotted.

If the seasonal argument is included, the winter and summer FDC lines will
be included on the plot as well.
}
\examples{

# read in Raven Hydrographs file
ff <- "C:/temp/model/output/run4_Hydrographs"
myhyd <- hyd.read(ff)

# extract flow for subbasin 24
flow.24 <- hyd.extract(subs="sub24",hyd=myhyd,period="2006-10-01/2010-10-01")

# create FDC plot
flowdurcurve.plot(flow.24,seasonal=T)

}
\seealso{
\code{\link{hyd.read}} for reading in the Hydrographs.csv file
\code{\link{hyd.extract}} for extracting basin flow information from a
hyd.read object

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{curve}
\keyword{diagnostics}
\keyword{duration}
\keyword{flow}
\keyword{plot}
