% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcings.read.R
\name{forcings.read}
\alias{forcings.read}
\title{Read in Raven ForcingFunctions file}
\usage{
forcings.read(ff = NA)
}
\arguments{
\item{ff}{full file path to the ForcingFunctions.csv file}
}
\value{
\item{wstor}{data frame from the file with standardized names}
\item{units}{vector corresponding to units of each res column}
}
\description{
forcings.read is used to read in the ForcingFunctions.csv file produced by
the modelling Framework Raven.
}
\details{
This function expects a full file path to the ForcingFunctions.csv file,
then reads in the file using read.csv. The main advantage of this functon is
renaming the columns to nicer names and extracting the units into something
much easier to read.

ff is the full file path of the ForcingFunctions.csv file. If the file is
located in the current working directory, then simply the name of the file
is sufficient.
}
\examples{

# read in sample forcings data
ff <- system.file("extdata","run1_ForcingFunctions.csv", package="RavenR")
myforcings <- forcings.read(ff)

# check data
head(myforcings$forcings)

}
\seealso{
\code{\link{hyd.read}} for reading in the Hydrographs.csv file

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{forcing}
\keyword{functions}
\keyword{read.csv}
