% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{get_rvi_options}
\alias{get_rvi_options}
\title{Provide known options for Raven rvi options}
\usage{
get_rvi_options()
}
\value{
string vector of recognized rvi_options
}
\description{
Provides a vector of recognized Raven rvi options for \code{\link{rvn_run}} to compare against.
}
\examples{
get_rvi_options()

}
\keyword{internal}
