% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{get_rvn_met_raven_mapping}
\alias{get_rvn_met_raven_mapping}
\title{Provide mappings for rvt functions}
\usage{
get_rvn_met_raven_mapping()
}
\value{
list of \code{rvn_met_raven_mapping}
}
\description{
Provides mapping of rvt_type and data_type that are used in the rvn_rvt functions.
}
\examples{
get_rvn_met_raven_mapping()

}
\keyword{internal}
