% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{get_rvt_met_mapping_weathercan}
\alias{get_rvt_met_mapping_weathercan}
\title{Provide mappings for weathercan to rvt functions}
\usage{
get_rvt_met_mapping_weathercan()
}
\value{
list of \code{rvt_met_mapping_weathercan}
}
\description{
Provides mapping of weathercan variable names to Raven that are used in the rvn_rvt functions.
}
\examples{
get_rvt_met_mapping_weathercan()

}
\keyword{internal}
