% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_info.R
\name{get_session_info}
\alias{get_session_info}
\title{Get list of session info with platform info and package list}
\usage{
get_session_info()
}
\description{
Get list of session info with platform info and package list
}
