% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceback.R
\name{get_traceback}
\alias{get_traceback}
\title{Get traceback as a list of strings}
\usage{
get_traceback(compact = TRUE)
}
\arguments{
\item{compact}{}
}
\value{

}
\description{
Get traceback as a list of strings
}
