% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceback.R
\name{get_warning_exception}
\alias{get_warning_exception}
\title{Get warning message from call stack (when warning itself is not available)}
\usage{
get_warning_exception()
}
\description{
Get warning message from call stack (when warning itself is not available)
}
