% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{ggplotColours}
\alias{ggplotColours}
\title{Generate colours consistent with ggplot defaults}
\usage{
ggplotColours(n = 6)
}
\arguments{
\item{n}{number of coours to generate}
}
\value{
vector of colours in hexadecimal format
}
\description{
Returns \code{n} colours consistent with default colours from \code{ggplot2}
}
\examples{
ggplotColours(4)
# returns c("#F8766D","#7CAE00","#00BFC4","#C77CFF")

}
