% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{\%notin\%}
\alias{\%notin\%}
\title{\%notin\% operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{values to be matched}

\item{table}{the values to be matched against}
}
\description{
Syntax for the opposite of \%in\%
}
\examples{
seq(1,5,1) \%notin\% seq(3,7,1)

}
