% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{hhmmss2dec}
\alias{hhmmss2dec}
\title{Convert hours, minutes, seconds to decimal hours}
\usage{
hhmmss2dec(x)
}
\arguments{
\item{x}{input as character, format HH:MM:SS}
}
\value{
{time in decimal hours}
}
\description{
Converts string format HH:MM:SS to decimal hours
}
\examples{
# return hour:minutes:seconds to decimal hours
hhmmss2dec("02:35:58")

}
