% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRUMap.plot.R
\name{HRUMap.plot}
\alias{HRUMap.plot}
\title{Plot continuous data using HRU shapefile}
\usage{
HRUMap.plot(shpfilename, HRUID_name = "ID", validHRUIDs, toplot,
  plottitle, legtitle, dmin = 1e+99, dmax = -1e+99, label = FALSE)
}
\value{
TRUE if completed successfully
}
\description{
Plots Raven custom output or other tabular data to HRU map
}
\details{
@param shpfilename filename of shapefile containing HRU polygons, with one column inidicating Raven HRU ID
 @param HRUID_name  name of shapefile column corresponding to HRU ID (default: "ID")
 @param validHRUIDs vector of valid modeled HRU IDs
 @param toplot  data frame of values to plot with row names equal to HRU IDs
 @param plottitle title of plot
 @param legtitle legend title for plot
 @param dmin data range minimum for determining color scale (optional)
 @param dmax data range maximum for determining color scale (optional)

does not currently support discrete data such as land use
}
\examples{
HRUdata<-HRUMap.prep.custom("SNOW_Monthly_Average_ByHRU.csv","2007-02-01")
HRUMap.plot("HRUs.shp",HRUdata$validHRUIDs,toplot=HRUdata$toplot,plottitle=HRUdata$plotttitle,legtitle=HRUdata$legtitle)

rvh<-rvh.read("example.rvh")
HRUdata<-HRUMap.prep.table(rvh$HRUtable$ID,rvh$HRUtable$Elevation,legtit="Elevation [m]")
HRUMap.plot("HRUs.shp",HRUdata$validHRUIDs,toplot=HRUdata$toplot,plottitle=HRUdata$plotttitle,legtitle=HRUdata$legtitle)

}
\seealso{
\code{\link{HRUMap.prep.table}} \code{\link{HRUMap.prep.custom}}
See also the \href{http://raven.uwaterloo.ca/}{Raven web site}
}
\author{
James R. Craig, University of Waterloo
}
