% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRUMap.plot.R
\name{HRUMap.prep.custom}
\alias{HRUMap.prep.custom}
\title{Prepare data for plotting from Raven custom output file}
\usage{
HRUMap.prep.custom(custfilename, date_toplot)
}
\arguments{
\item{custfilename}{Raven custom output file name (e.g., "SNOW_Monthly_Average_ByHRU.csv")}

\item{date_toplot}{date to plot aggregate results for "2007-02-01" (first of month for monthly, first of year for yearly)}
}
\value{
\item{validHRUIDs}{a vector of HRU ids which are present in the custom output file}
 \item{toplot}{a dataframe of data linked to each HRU, where rownames are HRU IDs}
 \item{legtitle}{a proposed legend title generated from the custom ouptut filename}
 \item{plottitle}{a proposed plot title generated from the custom ouptut filename}
}
\description{
This routine takes a custom output filename generated by the Raven modelling software
and extracts a single timestamp's worth of data in preparation for plotting using the
\code{\link{HRUMap.plot}} function. It also prepares a default legend and plot title
from the custom output file name
}
\examples{

\dontrun{
HRUdata<-HRUMap.prep.custom("SNOW_Monthly_Average_ByHRU.csv","2007-02-01")
HRUMap.plot("HRUs.shp",HRUdata$validHRUIDs,toplot=HRUdata$toplot,plottitle=HRUdata$plotttitle,legtitle=HRUdata$legtitle)
}
}
\seealso{
\code{\link{HRUMap.plot}}; See also the \href{http://raven.uwaterloo.ca/}{Raven web site}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{HRUs}
\keyword{Map}
\keyword{Plot}
\keyword{Raven}
