% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRUMap.plot.R
\name{HRUMap.prep.table}
\alias{HRUMap.prep.table}
\title{HRU plot preparation from table}
\usage{
HRUMap.prep.table(HRUIDs, DataVec, legtit = NULL, plottit = NULL)
}
\arguments{
\item{DataVec}{a vector or dataframe column of corresponding real-valued continuous data to plot}

\item{HRUVec}{a vector or dataframe column of HRU IDs}
}
\value{
\item{validHRUIDs}{a vector of HRU ids which are present in the shapefile}
 \item{toplot}{a dataframe of data linked to each HRU, where rownames are HRU IDs}
 \item{legtitle}{a proposed legend title}
 \item{plottitle}{a proposed plot title generated from the custom ouptut filename}
}
\description{
Prepares data for plotting from arbitrary data table
}
\seealso{
\code{\link{HRUMap.plot}}
See also the \href{http://raven.uwaterloo.ca/}{Raven web site}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{HRUs}
\keyword{Map}
\keyword{Plot}
\keyword{Raven}
