% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd.dyGraphs.R
\name{hyd.dyGraphs}
\alias{hyd.dyGraphs}
\title{Read in Raven Hydrograph file}
\usage{
hyd.dyGraphs(hy, timezone = "UTC", basins = "", figheight = 400)
}
\arguments{
\item{hy}{hydrograph data structure generated by hyd.read() routine}

\item{timezone}{data timezone; defaults to UTC}

\item{basins}{list of subbasin names from hydrograph file. Each subbasin creates separate dygraph plots}

\item{figheight}{height of figure, in pixels}
}
\value{
a list of plot handles to dygraph plots

See also \href{http://raven.uwaterloo.ca/}{Raven page}
}
\description{
hyd.dyGraphs plots modeled vs observed hydrographs when supplied with hydrograph data 
structure read using hyd.read()
}
\examples{

# read in Raven Hydrographs file
hy<-hyd.read("c:/TEMP/Hydrographs.csv",tz="MST")

# view contents for subbasin 11 as dyGraph
res<-hyd.dyGraphs(hy,basins="sub11")

}
\keyword{Raven}
\keyword{hydrograph}
\keyword{plotting}
