% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd.extract.R
\name{hyd.extract}
\alias{hyd.extract}
\title{Extract function for Raven Hydrograph object}
\usage{
hyd.extract(subs = NA, hyd = NA, prd = NULL)
}
\arguments{
\item{subs}{column name for plotting/extracting}

\item{hyd}{full hydrograph data frame (including units) produced by hyd.read}

\item{prd}{time period for plotting, as string. See details}
}
\value{
\item{sim}{model simulation for specified column and period}
\item{obs}{observed data for specified column and period}
\item{inflow}{inflow simulation for specified column and period}
}
\description{
hyd.extract is used for extracting data from the Raven hydrograph object.
Works for objects from hyd.extract function (from the Hydrographs.csv file).
}
\details{
hyd.extract is used to extract the modelled and observed data from a Raven
hydrograph object by name reference. It is also easy to create plots of
modelled and observed data using this function. The simulated and observed
files are outputted regardless of whether a plot is created, for the
specified period.

The subs input is the name of the column desired for use; the most common
use of this will be for subbasin outflows, where the names will be of the
form "subXX", for example "sub24".

The hyd object is the full hydrograph object (hyd and units in one data
frame) created by the hyd.read function. Both the hyd and units are
required, since the units are placed onto the plots if one is created. This
is useful to at least see the units of the plotted variable, even if the
plot is later modified.

The prd input is used to specify a period for the plot and/or the data
output. The period should be specified as a string start and end date, of
the format "YYYY-MM-DD/YYYY-MM-DD", for example, "2006-10-01/2010-10-01". If
no period is supplied, the entire time series will be used.
}
\examples{

# read in hydrograph sample csv data from RavenR package
ff <- system.file("extdata","run1_Hydrographs.csv", package="RavenR")

# read in Raven Hydrographs file, store into myhyd
myhyd <- hyd.read(ff)

# no plot or observed data, specified period
flow.36 <- hyd.extract(subs="Sub36",myhyd)

# plot the simulated flow values
plot(flow.36$sim,col='red')
# plot observed values on same plot
lines(flow.36$obs,lty=5,col='black')

# example for precipitation
myprecip <- hyd.extract(subs="precip",hyd=myhyd)

}
\seealso{
\code{\link{hyd.read}} for reading in the Hydrographs.csv file and
creating the object required in this function.

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{extract}
\keyword{hydrograph}
\keyword{plot}
