% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd.plot.R
\name{hyd.plot}
\alias{hyd.plot}
\title{Plot Hydrograph}
\usage{
hyd.plot(sim = NULL, obs = NULL, inflow = NULL, precip = NULL,
  prd = NULL, winter.shading = T, range.mult.hyd = 1.5,
  range.mult.precip = 1.5, ylabel = "Flow [m3/s]", leg.pos = NULL,
  leg.box = NULL, zero.axis = T)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{inflow}{time series object of inflows to subbasin}

\item{precip}{time series object of precipitation}

\item{prd}{period to use in plotting}

\item{winter.shading}{optionally adds shading for winter months (default
TRUE)}

\item{range.mult.hyd}{range multiplier for max value in hydrograph (default 1.5)}

\item{range.mult.precip}{range multiplier for max value in precipitation plot (default 1.5)}

\item{ylabel}{text label for y-axis of the plot (default 'Flow [m3/s]')}

\item{leg.pos}{string specifying legend placement on plot}

\item{leg.box}{boolean on whether to put legend in an opaque box}

\item{zero.axis}{fixes the y axis to start exactly at zero (default TRUE)}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
hyd.plot creates a hydrograph plot for the supplied flow series, or
equivalently a stage plot for reservoir stages.
}
\details{
This function creates a hydrograph plot using the supplied time series; any
series not supplied will not be plotted. If the precip time series is
supplied, the secondary y axis will be used to plot the precip time series.

The function assumes that the supplied time series have the same length and
duration in time. If this is not true, then the defined period or period
calculated from the first available flow series will be used to determine
the plotting limits in time. If the data is used directly from Raven output,
this is not a concern. The supplied time series should be in xts format,
which again can be obtained directly by using the hyd.extract function.

The winter.shading argument will add a transparent cyan shading for the
December 1st to March 31st period in each year that is plotted.

The range.mult arguments (range.mult.hyd, range.mult.precip) will increase the maximum value that is plotted in
the flows and the precip values. This is useful in preventing overlap if
precip is also plotted (i.e. with precip as well, range.mult=1.5 works
well). This value should not be less than 1.0, otherwise the values will be
cutoff in the plot. A value of 2 in each argument will ensure there is no overlap
in the hydrographs and precipitation plots.

ylabel is the label on the y axis, defined using y.lab in the plot function.
This defaults to 'Flow [m3/s]' intended for plotting hydrographs.

leg.pos is the position for the legend to be placed, e.g. 'topleft',
'right', etc., and is consistent with the legend function options. If this
is left null, the function will place it either topleft or left, depending
on whether precip is also plotted (i.e. left if precip added, topleft
otherwise).

leg.box is a boolean for whether to put the legend in an opaque. white box
or not. If left as NULL, the function will automatically not use a white box
and leave the background of the legend transparent.

zero.axis can be used to set the min value of the y axis (or axes if precip
also plotted) to zero. Note that by default, R will plot the values with a
slight buffer for presentation. A warning that if this option is set to
TRUE, the minimum value is set to zero without checking if any flow values
are less than zero. This option should not be used for stage plotting, since
most reservoirs are not in the range of 'zero' stage for normal operations,
since stage is reported to elevation and not to stage bottom, typically.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(hydrograph.data)
sim <- hydrograph.data$hyd$Sub36
obs <- hydrograph.data$hyd$Sub36_obs
precip <- hydrograph.data$hyd$precip

# create a nice hydrograph
hyd.plot(sim,obs,zero.axis=F)

# create a hydrograph with precip as well;
## range.mult=1.5 by default, leaves some overlap in plot axes
hyd.plot(sim,obs,precip=precip)

# hydrograph with no overlap, range.mult = 2
hyd.plot(sim,obs,range.mult.hyd=2,range.mult.precip=2,precip=precip)

# create a hydrograph with precip as well for a specific subperiod
prd <- "2003-10-01/2004-10-01"
hyd.plot(sim,obs,precip=precip,prd=prd)

}
\seealso{
\code{\link{flow.spaghetti}} to create a spaghetti plot of annual
flow series

\code{\link{hyd.extract}} to extract time series from Raven objects

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{flow}
\keyword{hydrograph}
