% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd.read.R
\name{hyd.read}
\alias{hyd.read}
\title{Read in Raven Hydrograph file}
\usage{
hyd.read(ff = NA, tzone = "")
}
\arguments{
\item{ff}{full file path to the Hydrographs.csv file}

\item{tzone}{string indicating the timezone of the data in ff}
}
\value{
\item{hyd}{data frame from the file with standardized names}
}
\description{
hyd.read is used to read in the Hydrographs.csv file produced by the
modelling Framework Raven.
}
\details{
This function expects a full file path to the Hydrographs.csv file, then
reads in the file using read.csv. The main advantage of this function is
renaming the columns to nicer names and extracting the units into something
much easier to read.

This function is also built to support the hyd.extract function, which uses
the object created here for extracting by reference to the columns named
here, for example sub24.

ff is the full file path of the Hydrographs.csv file. If the file is located
in the current working directory, then simply the name of the file is
sufficient.

tzone is a string indicating the timezone of the supplied Hydrographs file. The
timezone provided is coded into the resulting hyd data frame using the as.POSIXct
function. If no timezone is provided, this is left as an empty string, and is
determined by the function as the current time zone.
}
\examples{

# read in hydrograph sample csv data from RavenR package
ff <- system.file("extdata","run1_Hydrographs.csv", package="RavenR")

# read in Raven Hydrographs file, store into myhyd
myhyd <- hyd.read(ff)

# view contents
head(myhyd$hyd)
myhyd$units

}
\seealso{
\code{\link{hyd.extract}} for extraction tools related to the
hyd.read output file

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{hydrograph}
\keyword{read.csv}
