% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.vbias.R
\name{monthly.vbias}
\alias{monthly.vbias}
\title{Monthly Volume Bias}
\usage{
monthly.vbias(sim, obs, rplot = T, add.line = T, normalize = T,
  add.labels = T)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{rplot}{boolean whether to generate plot (default TRUE)}

\item{add.line}{optionally adds a horizontal line to the plot for reference
(default TRUE)}

\item{normalize}{option to normalize the biases and report as percent error
(default TRUE)}

\item{add.labels}{optionally adds labels for early peak/late peaks on right
side axis (default TRUE)}
}
\value{
\item{mvbias}{monthly volume biases}
}
\description{
monthly.vbias creates a plot of the monthly volume biases in the simulated
flow series.
}
\details{
This function calculates the monthly volume biases and optionally creates a
plot of them. The monthly volume biases are averaged across all years of
data. If normalized, the biases are calculated as:

(Vi_sim - Vi_obs)/Vi_obs*100

to be expressed as a percent error.

The sim and obs should be of time series (xts) format and are assumed to be
of the same length and time period. The flow series are assumed to be daily
flows with units of m3/s. Note that a plot title is purposely omitted in
order to allow the automatic generation of plot titles.

The add.labels will add the labels of 'overestimated' and 'underestimated'
to the right hand side axis if set to TRUE. This is useful in interpreting
the plots. Note that the biases are calculated as sim_Volume - obs_Volume,
which means that negative values mean the volume is underestimated, and
positive values mean the volume is overestimated.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(hydrograph.data)
sim <- hydrograph.data$hyd$Sub36
obs <- hydrograph.data$hyd$Sub36_obs

# check the monthly volume bias; normalizes by default
monthly.vbias(sim,obs)

# check unnormalzied monthly volume biases; see the larger volumes in certain periods
monthly.vbias(sim,obs,normalize = F)

}
\seealso{
\code{\link{annual.volume}} to create a scatterplot of annual flow
volumes

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{bias}
\keyword{diagnostics}
\keyword{monthly}
\keyword{volume}
