% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos.names.R
\name{mos.names}
\alias{mos.names}
\title{Months in the Year vector}
\usage{
mos.names(short = T)
}
\arguments{
\item{short}{boolean to return shortened form of months}
}
\value{
\item{int}{number of days between the two days}
}
\description{
mos.names is used to return a string vector of months in the year
}
\examples{

months_of_the_year <- mos.names
months_of_the_year

mos.names(F)

}
\seealso{
\code{\link{num.days.month}} for calculating the number of days in a
month
}
\keyword{months}
\keyword{year}
