% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.days.month.R
\name{num.days.month}
\alias{num.days.month}
\title{Number of Days in Month}
\usage{
num.days.month(date)
}
\arguments{
\item{date}{object in date format}
}
\value{
\item{int}{number of days in the month}
}
\description{
num.days.month is used to calculate the number of days in the month; handles
leap years
}
\examples{

num.days.month(as.Date("2016-02-05"))
# 29

num.days.month(as.Date("2017-01-17"))
# 31

}
\seealso{
\code{\link{num.days}} for calculating the number of days between
two dates

See original code on post in Stack Overflow
\href{http://stackoverflow.com/questions/6243088/find-out-the-number-of-days-of-a-month-in-rfind}{
the number of days in a month}
}
\keyword{days}
\keyword{month}
\keyword{number}
