% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsfile.create.R
\name{obsweights.gen}
\alias{obsweights.gen}
\title{Create weights time series for calibration/diagnostic evaluation}
\usage{
obsweights.gen(ts, criterion = "BETWEEN", startdate = "1785-10-05",
  enddate = "2500-01-01")
}
\arguments{
\item{ts}{xts time series}

\item{criterion}{criterion used to determine weighted vs. non-weighted dates
one of 'BEFORE', 'AFTER', 'BETWEEN', 'BETWEEN_CYCLIC'}

\item{startdate}{Date indicating start of time period (for 'BETWEEN' or
'AFTER' criterion) or start of annual cyclic period (for 'BETWEEN_CYCLIC'). In
the latter case, only the julian day of the startdate matters.}

\item{enddate}{Date indicating end of time period (for 'BETWEEN' or
'BEFORE' criterion) or end of annual cyclic period (for 'BETWEEN_CYCLIC'). In
the latter case, only the julian day of the enddate matters.}
}
\value{
nothing; produces file only
}
\description{
Creates an observation data weights time series based upon dates stored in
an xts time series and criterion given by the user
}
\details{
for criterion ='BEFORE', all timestamps prior to the enddate have a
weight of 1, 0 otherwise
for criterion ='AFTER', all timestamps after the startdate have a
weight of 1, 0 otherwise
for criterion ='BETWEEN', all timestamps after the startdate and before
the enddate have a weight of 1, 0 otherwise
for criterion ='BETWEEN_CYCLIC', all julian days after the startdate and before
the enddate have a weight of 1, 0 otherwise; if startdate is more than enddate,
then the opposite is true, e.g, for startdate="2002-11-01" and enddate "2002-01-31",
only November, December and January timestamps have a weight of 1
}
\examples{
  mydata<-read.csv("JohnCreek.csv")
  flows<-timeseries.infill(mydata)
  obsfile.create("JohnCreek.rvt", flows, 13,typestr="HYDROGRAPH")

  # just weight March-October flows :
  wts<-obsweights.gen(flows,criterion="BETWEEN_CYCLIC",startdate="2000-03-01",enddate="2000-11-01")

  # and only after March 2003:
  wts<-obsweights.gen(flows,criterion="AFTER",startdate="2003-03-01")
  wts2=wts2*wts # product merges weights

  obsweightsfile.create("JohnCreekWts.rvt",wts2,13,typestr="HYDROGRAPH")

}
\seealso{
\code{\link{obsweightsfile.create}}
See also the \href{http://raven.uwaterloo.ca/}{Raven website}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Raven}
\keyword{observations}
