% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsfile.create.R
\name{obsweightsfile.create}
\alias{obsweightsfile.create}
\title{Create Raven observation data weight (rvt) file}
\usage{
obsweightsfile.create(filename, wts, SBID, typestr = "HYDROGRAPH")
}
\arguments{
\item{filename}{observation data file to be created, with .rvt extension}

\item{wts}{xts time series with single data column of observation weights}

\item{SBID}{Subbasin ID for hydrographs and reservoir stages or HRU ID for observations of state variables (e.g., snow depth)}

\item{typestr}{Raven-recognized data type string: 'HYDROGRAPH', 'RESERVOIR_STAGE', 'RESERVOIR_INFLOW', 'RESERVOIR_NET_INFLOW',
or a state variable name (e.g., SOIL[0] or SNOW)}
}
\value{
nothing; produces file only
}
\description{
Creates an observation data weights file named filename from a continuous (gap-free) xts time series ts
with values between 0 and 1
}
\examples{
  mydata<-read.csv("JohnCreek.csv")
  flows<-timeseries.infill(mydata)
  obsfile.create("JohnCreek.rvt", flows, 13,typestr="HYDROGRAPH")

  # just weight March-October flows :
  wts<-obsweights.gen(flows,criterion="BETWEEN_CYCLIC",startdate="2000-03-01",enddate="2000-11-01")

  # and only after March 2003:
  wts<-obsweights.gen(flows,criterion="AFTER",startdate="2003-03-01")
  wts2=wts2*wts # product merges weights

  obsweightsfile.create("JohnCreekWts.rvt",wts2,13,typestr="HYDROGRAPH")

}
\seealso{
\code{\link{obsfile.create}} \code{\link{obsweights.gen}}
See also the \href{http://raven.uwaterloo.ca/}{Raven website}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Raven}
\keyword{observations}
