% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ost.params.compare.R
\name{Ost.params.compare}
\alias{Ost.params.compare}
\title{Compare parameters between Ostrich jobs}
\usage{
Ost.params.compare(ffs = NULL, num.metrics = NA, find.dev = F,
  num.sd = 1)
}
\arguments{
\item{ffs}{vector of Ostrich data files to compare}

\item{num.metrics}{number of fit metrics used in Ostrich; will remove the
metrics if supplied}

\item{find.dev}{boolean for finding parameter with large deviations between
jobs}

\item{num.sd}{number of standard deviations to define a large deviation
between jobs; default 1}
}
\value{
\item{Ost.paramcompare}{data frame of parameter values and
statistics} \item{dev.ind}{vector of indices showing parameter with some
variation between jobs}
}
\description{
Ost.params.compare is used to compare parameter values between different
Ostrich jobs.
}
\details{
Ost.params.compare is used to compare parameter values between different
Ostrich jobs of the same format (the number of metrics and parameter should
be the same between each file). This function is useful in determining which
parameters changed values by a large margin between two or more different
calibration jobs (based on the optimal objective function trial). In order
to compare the ensemble of equifinal parameter values between different
jobs, extra analysis for data returned from the Ost.bestparams function
should be done.

ffs is the vector of files to be read in and compared, which can be relative
or absolute pathways. The function will stop if any of these cannot be
found. It should be emphasized that the format between each of these files
needs to be the same in terms of number of metrics and parameters (i.e.
number of columns). The number of rows (calibration runs) can be different,
since this function will only extract the best run to examine.

num.metrics refers to the number of fit metrics used in the Ostrich
calibration, of which there should be at least one. If this value is
supplied, the function will remove the metric columns from the returned
data.

find.dev will return the indices of columns from the returned data that show
a large variation. Note that if this boolean is true, the returned object is
not just the dataset but a list of objects (see example below).

num.sd indicates the number of standard deviations used in the definition of
a 'large' variation between parameter values. To assume a normal
distribution and 95 percent interval, one can use 1.96 here.
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{
ff1 <- "1_OstModel0.txt"
ff2 <- "2_OstModel0.txt"
ff3 <- "3_OstModel0.txt"
ffm <- c(ff1,ff2,ff3)

mydata <- Ost.params.compare(ffm,find.dev=F)
head(mydata)

mydata <- Ost.params.compare(ffm,find.dev=F)
mydata$dev.ind  # indices where some job had largely different value(s)
head(mydata$Ost.paramcompare)  # data frame of parameter values and statistics

# show parameter values with deviations
mydata$Ost.paramcompare[,mydata$dev.ind]
}

}
\seealso{
\code{\link{Ost.read}} for reading in Ostrich files

\code{\link{Ost.plot}} for creating various plots with Ostrich data

\code{\link{Ost.bestparams}} for obtaining the best parameter sets from
Ostrich data

Download Ostrich at
\href{http://www.eng.buffalo.edu/~lsmatott/Ostrich/OstrichMain.html}{Shawn
Matott's web page}

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for other software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Ostrich}
\keyword{calibration}
\keyword{parameter}
