% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ost.plot.R
\name{Ost.plot}
\alias{Ost.plot}
\title{Plot data from Ostrich output}
\usage{
Ost.plot(Ost.data = NA, plot.type = "oft", return.data = F,
  logy = F, rangey = NA, num.metrics = 1, param.index = 1)
}
\arguments{
\item{Ost.data}{data object from Ost.read function (optional)}

\item{plot.type}{the type of plot to generate}

\item{return.data}{whether to return the Ostrich data used in plotting}

\item{logy}{boolean to plot the y axis on a log scale}

\item{rangey}{range of the y axis; optional}

\item{num.metrics}{number of fit metrics used in Ostrich}

\item{param.index}{index of parameter to plot for parameter sensitivity
plots}
}
\value{
\item{Ost.dd}{Ostrich data used in plotting; returned only if
return.data is TRUE. Else function returns TRUE if the function executed
properly}
}
\description{
Ost.plot is used to plot data from calibrations performed using Ostrich.
}
\details{
The Ost.plot is used to create various types of plots using the data from an
Ostrich calibration run. For simplicity only the 'OstModel0.txt' file or
data is used; this can be passed as a data object from Ost.read, or if NULL
the function will look for the OstModel0.txt file in the current directory
and use that.

The plot.type can be either 'oft', 'boft', or 'psa', which refers to
objective function vs trials, best objective function vs trials, and
parameter sensitivity analysis for a single parameter plot.

logy is used to plot the y axis logarithmically; note that this only applies
for the plot.type oft or boft, not for the psa plots.

rangey only applies to the oft plot type.

num.metrics refers to the number of fit metrics used in the Ostrich
calibration; note that there should be at least one fit metric used in each
calibration run.

param.index refers to the index of the parameter you wish to plot for the
psa plot type. Note that the index is set for the parameters specifically,
not the column index in the OstModel0.txt file. For example, setting
param.index = 1 will create a plot of the first parameter in the
OstModel0.txt file.
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{

# locate sample OstModel0.txt file in RavenR package
ff <- system.file("extdata","OstModel0.txt", package="RavenR")

# read in file
mydata <- Ost.read(ff)

# simple plotting of the oft plot type
Ost.plot(Ost.data=mydata)

# plot of the best objective function vs trials
Ost.plot(mydata,plot.type='boft')

# sensitivty analysis of the 5th parameter
Ost.plot(mydata,plot.type='psa',num.metrics=4,param.index=5)
}

}
\seealso{
\code{\link{hyd.read}} for handling Raven hydrograph files

Download Ostrich at
\href{http://www.eng.buffalo.edu/~lsmatott/Ostrich/OstrichMain.html}{Shawn
Matott's web page}

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for other software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Ostrich}
\keyword{calibration}
