% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator

See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.}
\usage{
lhs \%>\% rhs
}
\description{
Pipe operator

See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
