% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.subbasinNetwork.R
\name{plot.subbasinNetwork}
\alias{plot.subbasinNetwork}
\title{Plot Subbasin Network}
\usage{
\method{plot}{subbasinNetwork}(SBtable)
}
\arguments{
\item{SBnetwork}{SBnetwork output from rvh.read function}
}
\value{
\item{net}{igraph net object of watershed}
}
\description{
plot.subbasinNetwork creates a network plot of the subbasins using the igraph package
}
\details{
Plots subbasin network connectivity as specified in a Raven .rvh file. Centroids of
basins in lat-long are used to orient the network, so it is geographically consistent.
Areas are proportional to subbasin total area. Labels refer to subbasin ID.

Depends: igraph library.

Function contributed by Dr. James Craig
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{
# Using a valid Raven model .rvh file in the working directory,
# read the file into the rvh object
rvh<-rvh.read("WatershedX.rvh")
plot
plot.subbasinNetwork(rvh$SBtable)
}

}
\seealso{
\code{\link{rvh.read}} for reading in rvh files

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{HRU}
\keyword{Raven}
\keyword{igraph}
\keyword{network}
\keyword{plot}
\keyword{rvh}
\keyword{subbasin}
\keyword{watershed}
