% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RavenR-package.R
\docType{package}
\name{RavenR-package}
\alias{RavenR}
\alias{RavenR-package}
\title{RavenR}
\description{
Contains functions to perform pre- and post-processing of RavenR files.
}
\details{
This package provides a number of types of useful functions, including:
\itemize{
\item{reading in \strong{Raven output files} (including ForcingFunctions, Hydrographs, etc.)}
\item{processing some file types into \strong{Raven format}}
\item{flow-based \strong{diagnostics} and plot functions}
\item{other useful plotting functions (e.g. rvi process plot, spaghetti plot, dyGraphs, network plots)}
\item{useful utilities for time series and hydrologic functions (e.g. rvn_apply_wyearly)}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rchlumsk/RavenR}
  \item Report bugs at \url{https://github.com/rchlumsk/RavenR/issues}
}


The package on \href{https://cran.r-project.org/package=RavenR}{CRAN} and
the development version on \href{https://github.com/rchlumsk/RavenR}{GitHub}.
See also \href{https://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{https://raven.uwaterloo.ca/}{Raven webpage}.
}
\author{
\strong{Maintainer}: Robert Chlumsky \email{rchlumsk@uwaterloo.ca} (\href{https://orcid.org/0000-0002-1303-5064}{ORCID})

Authors:
\itemize{
  \item James Craig \email{jrcraig@uwaterloo.ca} (\href{https://orcid.org/0000-0003-2715-7166}{ORCID}) [contributor]
  \item Leland Scantlebury \email{lscantlebury@uwaterloo.ca} [contributor]
  \item Simon Lin \email{simon.lin@uwaterloo.ca} [contributor]
  \item Sarah Grass \email{sgrass@nhcweb.com} [contributor]
  \item Genevieve Brown \email{genevieve.wbrown@gmail.com} [contributor]
  \item Rezgar Arabzadeh \email{rarabzad@uwaterloo.ca} [contributor]
}

}
\keyword{internal}
