% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.extract.R
\name{res.extract}
\alias{res.extract}
\title{Extract function for Raven Reservoir object}
\usage{
res.extract(subs = NA, res = NA, prd = NULL)
}
\arguments{
\item{subs}{column name for plotting/extracting}

\item{res}{full reservoir data frame (including units) produced by res.read}

\item{prd}{time period for plotting, as string. See details}
}
\value{
\item{sim}{model simulation for specified column and period}
\item{obs}{observed data for specified column and period}
\item{inflow}{inflow simulation for specified column and period}
}
\description{
res.extract is used for extracting data from the Raven reservoir object.
Works for objects from res.extract function (from the ReservoirStages.csv
file).
}
\details{
res.extract is used to extract the modelled and observed data from a Raven
reservoir object by name reference. It is also easy to create plots of
modelled and observed data using this function. The simulated and observed
files are outputted regardless of whether a plot is created, for the
specified period.

The subs input is the name of the column desired for use; the most common
use of this will be for subbasins, where the names will be of the form
"subXX", for example "sub24".

The res object is the full reservoir object (res and units in one data
frame) created by the res.read function. Both the res and units are
required, since the units are placed onto the plots if one is created. This
is useful to at least see the units of the plotted variable, even if the
plot is later modified.

The prd input is used to specify a period for the plot and/or the data
output. The period should be specified as a string start and end date, of
the format "YYYY-MM-DD/YYYY-MM-DD", for example, "2006-10-01/2010-10-01". If
no period is supplied, the entire time series will be used.
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{
# create full file path
ff <- "C:/temp/model/output/run4_Reservoirs.csv")

# read in Raven Reservoirs file, store into myres
myres <- res.read(ff)

# no plot or observed data, specified period
stage.24 <- res.extract(subs="sub24",res=myres,prd="2006-10-01/2010-10-01")

# plot the simulated stage values
plot(stage.24$sim,col='red')
# plot observed values on same plot
lines(stage.24$obs,lty=5)

# example for precipitation
myplot <- res.extract(subs="precip",res=myres)
}

}
\seealso{
\code{\link{res.read}} for reading in the Reservoirs.csv file and
creating the object required in this function. \code{\link{res.plot}} for
plotting the extracted stage time series

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{extract}
\keyword{plot}
\keyword{reservoir}
