% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.plot.R
\name{res.plot}
\alias{res.plot}
\title{Plot Reservoir Stage}
\usage{
res.plot(sim = NULL, obs = NULL, inflow = NULL, precip = NULL,
  prd = NULL, winter.shading = T, range.mult = 1,
  ylabel = "Stage [m]", leg.pos = NULL, leg.box = NULL,
  zero.axis = T)
}
\arguments{
\item{sim}{time series object of simulated stage}

\item{obs}{time series object of observed stage}

\item{inflow}{time series object of inflow to subbasin}

\item{precip}{time series object of precipitation}

\item{prd}{period to use in plotting}

\item{winter.shading}{optionally adds shading for winter months (default
TRUE)}

\item{range.mult}{range multiplier for max value in stage and precip
(default 1)}

\item{ylabel}{label on the y axis (default "Stage [m]")}

\item{leg.pos}{string specifying legend placement on plot}

\item{leg.box}{boolean on whether to put legend in an opaque box}

\item{zero.axis}{fixes the y axis to start exactly at zero (default TRUE)}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
res.plot creates a reservoir stage plot for the supplied flow series, or
equivalently a stage plot for reservoir stages.
}
\details{
This function creates a reservoir stage plot using the supplied time series;
any series not supplied will not be plotted. If the precip time series is
supplied, the secondary y axis will be used to plot the precip time series.

The function assumes that the supplied time series have the same length and
duration in time. If this is not true, then the defined period or period
calculated from the first available flow series will be used to determine
the plotting limits in time. If the data is used directly from Raven output,
this is not a concern. The supplied time series should be in xts format,
which again can be obtained directly by using the res.extract function.

The winter.shading argument will add a transparent cyan shading for the
December 1st to March 31st period in each year that is plotted.

The range.mult argument will increase the maximum value that is plotted in
the stage and the precip values. This is useful in preventing overlap if
precip is also plotted (i.e. with precip as well, range.mult=1.5 works
well). This value should not be less than 1.0, otherwise the values will be
cutoff in the plot.

ylabel is the label on the y axis, defined using y.lab in the plot function.
This defaults to 'Stage [m]' intended for plotting stage.

leg.pos is the position for the legend to be placed, e.g. 'topleft',
'right', etc., and is consistent with the legend function options. If this
is left null, the function will place it either topleft or left, depending
on whether precip is also plotted (i.e. left if precip added, topleft
otherwise).

leg.box is a boolean for whether to put the legend in an opaque. white box
or not. If left as NULL, the function will automatically not use a white box
and leave the background of the legend transparent.

zero.axis can be used to set the min value of the y axis (or axes if precip
also plotted) to zero. Note that by default, R will plot the values with a
slight buffer for presentation. A warning that if this option is set to
TRUE, the minimum value is set to zero without checking if any flow values
are less than zero. This option should not be used for stage plotting, since
most reservoirs are not in the range of 'zero' stage for normal operations,
since stage is reported to elevation and not to stage bottom, typically.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{
# create a nice reservoir stage plot
res.plot(sim,obs,zero.axis=F)

# create a reservoir stage plot with precip as well
res.plot(sim,obs,range.mult=1.5,precip=precip)

# create a reservoir stage plot with precip as well for a specific subperiod
prd <- "2003-10-01/2005-10-01"
res.plot(sim,obs,range.mult=1.5,precip=precip,prd=prd)
}

}
\seealso{
\code{\link{hyd.read}} for reading in the Hydrographs.csv file

\code{\link{res.extract}} to extract time series from Raven objects

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{reservoir}
\keyword{stage}
