% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvh.cleanHRUs.R
\name{rvh.cleanHRUs}
\alias{rvh.cleanHRUs}
\title{Clean HRU data table.}
\usage{
rvh.cleanHRUs(HRUtab, SBtab, area_tol = 0.001, merge = FALSE,
  elev_tol = 50, slope_tol = 4, aspect_tol = 20,
  ProtectedHRUList = c())
}
\arguments{
\item{HRUtab}{table of HRUs generated (typically) by \code{\link{rvh.read}}}

\item{SBtab}{table of Subbasins generated (typically) by \code{\link{rvh.read}}}

\item{area_tol}{percentage of watershed area beneath which HRUs should be removed (e.g.,
0.01 would indicate anything smaller than 1 percent of watershed extent should be removed)}

\item{merge}{TRUE if similar HRUs are to be merged (this can be slow for large models)}

\item{elev_tol}{elevation difference (in metres) considered similar. only used if merge=TRUE.}

\item{slope_tol}{slope difference (in degrees) considered similar. only used if merge=TRUE.}

\item{aspect_tol}{slope difference (in degrees) considered similar. only used if merge=TRUE.}

\item{ProtectedHRUList}{list of HRU IDs that are sacrosanct (not to be removed)}
}
\value{
revised HRU table as a dataframe
}
\description{
Takes \code{\link{rvh.read}}-generated HRUtable and SBTable and returns cleaned HRUtable
with (hopefully) fewer HRUs
}
\details{
rvh.clean removes HRUs in two ways:
    (1) it removes all HRUs smaller than the area_tol percentage of total area. Adjacent HRUs in the
    subbasin are expanded by the lost area to keep the same relative coverage.
    (2) it consolidates similar HRUs within the same subbasin (those with same land cover,
    vegetation, soil profile and similar slope, aspect, and elevation)
}
\examples{
  #read in rvh file, clean contents and write back to new file
  rvh<-rvh.read("example.rvh")
  rvh<-rvh.cleanHRUs(rvh$HRUtable,rvh$SBtable,merg=TRUE)
  rvh.overwrite("cleanmodel.rvh","example.rvh",rvh$HRUtable,rvh$SBtable)

}
\seealso{
\code{\link{rvh.read}} for the function used to read in the HRU and SubBasin data
See also the \href{http://raven.uwaterloo.ca/}{Raven web site}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{HRUs}
\keyword{Raven}
\keyword{clean}
\keyword{merge}
\keyword{rvh}
