% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvh.read.R
\name{rvh.read}
\alias{rvh.read}
\title{Read Raven RVH File}
\usage{
rvh.read(filename)
}
\arguments{
\item{filename}{file name of the rvh file}
}
\value{
\item{SBtable}{data table of subbasin properties and statistics}

\item{HRUtable}{data table of HRU properties and statistics}

\item{SBnetwork}{igraph object for creating a network plot with
the plot.subbasinnetwork function}
}
\description{
rvh.read reads in a given rvh file and generates a data frame of useful
summary statistics
}
\details{
This function reads in a given Raven rvh file, and calculates a number
of useful statistics, including: total upstream area, dominant land cover/
vegetation info, and area-weighted aspect/slope/lat/long.

Function contributed by Dr. James Craig
}
\examples{

# read in rvh file from Alouette tutorial
rvh <- rvh.read('Alouette.rvh')

}
\seealso{
\code{\link{plot.subbasinNetwork}} for creating a network plot of the watershed

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{HRU}
\keyword{Raven}
\keyword{rvh}
\keyword{subbasin}
\keyword{watershed}
