% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvh.write.R
\name{rvh.write}
\alias{rvh.write}
\alias{rvh.overwrite}
\title{Write/Overwrite Raven .rvh (watershed discretization) file.}
\usage{
rvh.write(filename, SBtable, HRUtable)

rvh.overwrite(orig_file, filename, SBtable, HRUtable)
}
\arguments{
\item{filename}{Name of .rvh file, with extension}

\item{SBtable}{Valid subbasin dataframe with required columns "SBID", "Name", "Downstream_ID", "Profile",
"ReachLength", and "Gauged". Can have additional columns.}

\item{HRUtable}{Valid HRU dataframe with required columns 'ID', 'Area', 'Elevation', 'Latitude', 'Longitude',
'SBID', 'LandUse', 'Vegetation', 'SoilProfile', 'Terrain', 'Aquifer', 'Slope', and 'Aspect'. Can have additional columns.}

\item{orig_file}{original .rvh file, if overwriting. Must be different from  that specified by filename.}
}
\value{
none, generates/revises .rvh file only
}
\description{
Given an HRU and SubBasin dataframe, writes to the specified .rvh file or, in the case of rvh.overwrite,
just overwrites the :SubBasins-:EndSubBasins and :HRUs-:EndHRUs blocks, retaining all other content.
}
\details{
rvh.write is typically used to create a 'clean' .rvh file. rvh.overwrite is usually used
  after reading an original .rvh file and processing the HRU and SubBasin tables, using (e.g., rvh.clean)
}
\section{Functions}{
\itemize{
\item \code{rvh.overwrite}: Overwrite contents of original .rvh file
}}

\examples{
  #read in rvh file, clean contents and write back to new file
  rvh<-rvh.read("model.rvh")
  rvh<-rvh.cleanHRUs(rvh$HRUtable,rvh$SBtable)
  rvh.overwrite("cleanmodel.rvh","model.rvh",rvh$HRUtable,rvh$SBtable)
  rvh.write("nostrings.rvh",rvh$HRUtable,rvh$SBtable)
}
\seealso{
\code{\link{rvh.read}} for the function used to read in the HRU and SubBasin data
\code{\link{rvh.cleanHRUs}} for the function used to process HRU dataframes
See also the \href{http://raven.uwaterloo.ca/}{Raven web site}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{HRUs}
\keyword{Raven}
\keyword{rvh}
\keyword{write}
