% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvi.read.R
\name{rvi.read}
\alias{rvi.read}
\title{Read Raven .rvi (watershed discretization) file}
\usage{
rvi.read(filename)
}
\arguments{
\item{filename}{the name of the .rvi file (with .rvi extension included ), either relative
to the working directory or absolute.}
}
\value{
Returns a list including a lone item:
\item{HydProcTable}{a data table of hydrologic processes. Includes the following data columns:
process type, algorithm, 'from' compartment, 'to' compartment, and condition}
}
\description{
This routine reads in a valid Raven main input (.rvi) file and returns the
information about hydrological processes as a data table.
}
\details{
does not like comma-delimited tables with a trailing comma
}
\note{
does not like tabs in the .rvi file - it should be untabified first.
The .rvi file can have arbitrary contents outside of the :HydrologicProcesses-
:EndHydrologicProcesses block and :SubBasins-:EndSubBasins command blocks.
}
\examples{
 \dontrun{
  # sample workflow of rvh.read

  rvi <- rvi.read("example.rvi")

  # get number of Hydrologic processes
  numProcss <- nrow(rvi$HydProcTable)
}
}
\seealso{
rvi.connections
See also the \href{http://raven.uwaterloo.ca/}{Raven page}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Hydrologic}
\keyword{Processes}
\keyword{Raven}
\keyword{read}
\keyword{rvi}
