% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_annual_peak.R
\name{rvn_annual_peak}
\alias{rvn_annual_peak}
\title{Annual Peak Comparison}
\usage{
rvn_annual_peak(sim, obs, add_line = T, add_r2 = F, add_eqn = F, rplot = F)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{add_line}{optionally adds a 1:1 line to the plot for reference (default TRUE)}

\item{add_r2}{optionally computes the R2 and adds to plot (default FALSE)}

\item{rplot}{boolean whether to print the plot (default FALSE)}

\item{add_egn}{optionally adds the equation for a linear regression line through the origin (default FALSE)}
}
\value{
returns a list with peak data in a data frame, and a ggplot object
 \item{df_peak}{data frame of the calculated peaks}
 \item{p1}{ggplot object with plotted annual peaks}
}
\description{
rvn_annual_peak creates a plot of the annual observed and simulated peaks, based on the water year.
}
\details{
This function creates a scatterplot of the annual observed and simulated
peaks, calculated for each available water year of data (Oct 1st hardcoded)
within the two series provided. Note that the calculation uses the peak
magnitude of simulated and observed series in each water year, without
considering the timing of the events in each series.

The sim and obs should be of time series (xts) format and are assumed to be
of the same length and time period. The flow series are assumed to be daily
flows with units of m3/s.

The R2 diagnostic is calculated for a fit with no intercept (in a perfect
fit the points are identical, and intercept is automatically zero).

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{
# load sample hydrograph data, two years worth of sim/obs
data(rvn_hydrograph_data)
sim <- rvn_hydrograph_data$hyd$Sub36
obs <- rvn_hydrograph_data$hyd$Sub36_obs

# create a plot of annual peaks with default options
peak1 <- rvn_annual_peak(sim, obs, add_line=T, add_r2=F)
peak1$df_peak
peak1$p1

# plot with r2 and axes to zero; store results
peak_df <- rvn_annual_peak(sim, obs, add_line=T,
add_r2=T, rplot=T)

}
\seealso{
\code{\link{rvn_annual_volume}} to create a scatterplot of annual flow
volumes \code{\link{rvn_annual_peak_event}} to consider the timing of peak
events.

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{annual}
\keyword{diagnostics}
\keyword{peak}
