% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_annual_quantiles.R
\name{rvn_annual_quantiles}
\alias{rvn_annual_quantiles}
\title{Calculates Yearly Median, Upper and Lower Quantiles of Flow}
\usage{
rvn_annual_quantiles(
  hgdata,
  minyear = NULL,
  maxyear = NULL,
  Qlower = 0.1,
  Qupper = 0.9,
  water_year = T
)
}
\arguments{
\item{hgdata}{Time series object of observed or simulated flows}

\item{minyear}{Integer of minimum year (optional)}

\item{maxyear}{Integer of maximum year (optional)}

\item{Qlower}{Decimal percentage of lower quantile value (default 0.1)}

\item{Qupper}{Decimal percentage of upper quantile value (default 0.9)#'}
}
\value{
\item{qdat}{Time series object of monthly median and quantile values}
}
\description{
Calculates Yearly Median, Upper and Lower Quantiles of Flow
}
\examples{
system.file("extdata","run1_Hydrographs.csv", package="RavenR") \%>\%
rvn_hyd_read(.) \%>\%
rvn_hyd_extract(subs="Sub36",.) ->
hyd_data

# Pull out a specific hydrograph
hgdata <- rvn_hydrograph_data$hyd$Sub36

# Calculate quantiles
qdat <- rvn_annual_quantiles(hgdata)
head(qdat)

}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
\keyword{diagnostic}
\keyword{flow}
\keyword{hydrograph}
\keyword{quantile}
\keyword{statistics}
