% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_annual_quantiles_plot.R
\name{rvn_annual_quantiles_plot}
\alias{rvn_annual_quantiles_plot}
\title{Plot of Yearly Median, Upper and Lower Quantiles of Flow}
\usage{
rvn_annual_quantiles_plot(
  qdat,
  mediancolor = "black",
  ribboncolor = "grey60",
  ribbonalpha = 0.5,
  addtoplot = NULL
)
}
\arguments{
\item{qdat}{Time series object generated by rvn_annual_quantiles()}

\item{mediancolor}{Color for the median line}

\item{ribboncolor}{Color for the lower/upper quantile ribbon}

\item{ribbonalpha}{Transparency of lower/upper quantile ribbon}

\item{addtoplot}{Existing ggplot object to which median line and quantile ribbon should be added}
}
\value{
Monthly Plot
}
\description{
Plot of Yearly Median, Upper and Lower Quantiles of Flow
}
\examples{
data(hydrograph.data)

# Pull out a specific hydrograph
hgdata <- rvn_hydrograph_data$hyd$Sub36_obs

# Calculate quantiles
qdat <- yearly_quantile(hgdata)

# Plot
p <- rvn_yearly_quantile_plot(qdat)
p  # To view

# Add a second hydrograph to compare
simdata <- rvn_hydrograph_data$hyd$Sub36
qdat_sim <- rvn_annual_quantiles(hgdata)

p <- rvn_annual_quantiles_plot(qdat_sim, mediancolor = 'red', ribboncolor = 'red', addtoplot = p)
p # To view
}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
