% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_apply_wyearly.R
\name{rvn_apply_wyearly}
\alias{rvn_apply_wyearly}
\title{Apply function for water year}
\usage{
rvn_apply_wyearly(x, FUN, ...)
}
\arguments{
\item{x}{xts vector to calculate FUN for}

\item{FUN}{the function to be applied}

\item{...}{optional arguments to FUN}
}
\description{
apply_wyearly calculates a function FUN for the periods defined by the water
year, similar to other functions of the form apply.<time period>, for
example apply.daily, apply.monthly, etc. This is a function especially
helpful to hydrology data or results. The assumed water year start is
October 1st.
}
\details{
The function assumes an October 1st start to the water year. If the data
supplied is from, e.g. 2008-10-01 to 2009-09-30, the function will use the
September 30th date as the final day to apply the function for the water
year.

Note that currently the water year calculation is done from Oct 2 to Oct 1st
(or September 30th) inclusive. For example in the Oct 1 2007 to Oct 1 2008
water year, the end date is shown as "2008-10-01", and the function will be
evaluated for the period 2007-10-02 - 2008-10-01, inclusive.

The apply_wyearly function applies to only one column at a time, otherwise
only the first column will be used, with a warning.
}
\examples{

# use sample forcing data (or use forcings_read to read in ForcingFunctions.csv)
data(forcing_data)
myforcings <- forcing_data

# apply mean as FUN to daily average temperature
apply_wyearly(myforcings$forcings$temp_daily_ave,mean,na.rm=T

# use apply_wyearly_cols to apply to multiple columns at once
apply_wyearly_cols(myforcings$forcings[,2:6],max,na.rm=T)

}
\seealso{
\code{\link{apply_wyearly_cols}} for applying this function to
multiple columns explicitly

See also \href{http://www.civil.uwaterloo.ca/jrcraig/software.html}{James R.
Craig's research page} for software downloads
}
\keyword{apply}
\keyword{water}
\keyword{year}
