% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvh_write.R
\name{rvn_blankHRUdf}
\alias{rvn_blankHRUdf}
\title{Generate Blank Raven HRU DataFrame}
\usage{
rvn_blankHRUdf(nhrus = 1, subbasinIDs = NULL)
}
\arguments{
\item{nhrus}{Number of HRUs, used to determine number of rows in table (default = 1)}

\item{subbasinIDs}{Subbasins that HRUs belong to (default = all equal 1)}
}
\value{
data.frame of blank HRU properties to be filled in by user
}
\description{
Used to generate a blank HRU table that can be filled in by the user. Compatible with
rvn_rvh_write().
}
\examples{
HRUtable <- rvh_blankHRUdf(nhrus = 3, subbasinIDs=c(1,1,2))
}
\author{
Leland Scantlebury
}
