% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvh_write.R
\name{rvn_blankSubBasindf}
\alias{rvn_blankSubBasindf}
\title{Generate Blank Raven SubBasin DataFrame}
\usage{
rvn_blankSubBasindf(nsubbasins = 1)
}
\arguments{
\item{nsubbasins}{Number of SubBasins in model, used to determine number of rows in table (default = 1)}
}
\value{
data.frame of blank SubBasin properties to be filled in by user
}
\description{
Generate Blank Raven SubBasin DataFrame
}
\examples{
SBtable <- rvn_blankSubBasindf(nsubbasins = 1)
}
