% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_calc_areas_rvh.R
\name{rvn_calc_areas_rvh}
\alias{rvn_calc_areas_rvh}
\title{Calculate areas from Raven .rvh file}
\usage{
rvn_calc_areas_rvh(ff, calc_da = T, write_file = F, iter_thresh = NA)
}
\arguments{
\item{ff}{full filepath to .rvh file}

\item{calc_da}{boolean to calculate drainage areas}

\item{write_file}{boolean to write data to file}

\item{iter_thresh}{value to overwrite default iteration threshold in drainage
area calculation}
}
\value{
\item{areas_rvh}{matrix with calculated areas for each subbasin}
}
\description{
calc_areas_rvh is used to calculate the total area and drainage area for
each subbasin, from a Raven-formatted .rvh file
}
\details{
This function reads in the .rvh file in Raven format, and firstly calculates
the total area of each subbasin by linking the IDs to the HRU section.

The second piece of functionality is to use the network connections in the
Subbasins section to calculate the drainage area of each subbasin. The basic
step by step implementation of the code is as follows: 1) for all headwater
basins, assign drainage_area = basin_area 2) iterate through all subbasins;
if the drainage area is not calculated, calculate it if all basins that feed
water to that basin have their drainage areas calculated, otherwise skip 3)
repeat step 2 until all drainage areas are no longer NA 4) break if the
iter_thresh is exceeded for the number of iterations of Step 2

This is a simple drainage area calculator that works for any .rvh
configuration, as long the connections are feasible.

Note that the reading of the .rvh file currently assumes that there are no
comments on lines between :SubBasin and :EndSubBasins, nor in the HRUs
block.
}
\examples{

# locate in RavenR rvh sample file
ff <- system.file("extdata","Nith.rvh", package="RavenR")

# basic, with defaults (calculates drainage area also)
myareas <- calc_areas_rvh(ff)

# avoid calculating drainage areas
myareas <- calc_areas_rvh(ff,calc_da=F)

# calculate drainage areas, adjust iteration threshold
myareas <- calc_areas_rvh(ff,iter_thresh=1000)

# write results to file
myareas <- calc_areas_rvh(ff,write_file=T)

}
\seealso{
\code{\link{forcings_read}} for analysis tools related to the
forcing functions

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{area}
\keyword{drainage}
\keyword{rvh}
