% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_flow_residuals.R
\name{rvn_flow_residuals}
\alias{rvn_flow_residuals}
\title{Scatterplot of model flows}
\usage{
rvn_flow_residuals(sim, obs, ma_smooth = 3, add.line = T, winter_shading = T)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{ma_smooth}{optional length of rolling average to smooth residuals with
(default 3)}

\item{add.line}{optionally adds a horizontal line to the plot for reference
(default TRUE)}

\item{winter_shading}{optionally adds a light blue shading to winter months
(default TRUE)}
}
\value{
\item{resids}{residual time series}
}
\description{
rvn_flow_residuals creates a residuals time series for flow values. Useful in
diagnotic analysis of model outputs.
}
\details{
This function creates a residuals time series plot for flow values, with the
option to smooth out the values using the rollmean function in zoo. The
winter months are optionally shaded in the time series; winter period is
defined as December 1st to March 31st.

The residuals are calculated as sim - obs.

The sim and obs should be of time series (xts) format. The flow series are
assumed to be daily flows with units of m3/s.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(rvn_hydrograph_data)
sim <- rvn_hydrograph_data$hyd$Sub36
obs <- rvn_hydrograph_data$hyd$Sub36_ob

# default with moving average smoothing shading of winter months
rvn_flow_residuals(sim,obs)

# plot with more smoothing than the default 3
rvn_flow_residuals(sim, obs, ma_smooth=10)

# turn off the smoothing and winter shading
rvn_flow_residuals(sim,obs,ma_smooth = 0, winter_shading = F)

}
\seealso{
\code{\link{flow_scatterplot}} to create a scatterplot of flow
values

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{diagnostics}
\keyword{flow}
\keyword{residual}
