% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_forcings_plot.R
\name{rvn_forcings_plot}
\alias{rvn_forcings_plot}
\title{Plots summary of watershed forcing functions}
\usage{
rvn_forcings_plot(forcings, prd = NULL)
}
\arguments{
\item{forcings}{forcings attribute from forcings.read function}

\item{prd}{(optional) time period over which the plots are generated}
}
\description{
rvn_forcings_plot generates a set of 5 plots (precip,temperature,PET,radiation,
and potential melt), which summarize the watershed-averaged forcings.
}
\details{
This function creates multiple plots from a ForcingFunctions.csv file
structure generating using RavenR's forcings.read function
}
\examples{

# read in sample forcings data
data("rvn_forcing_data")
fdata <- rvn_forcing_data$forcings

# plot forcings data
rvn_forcings_plot(fdata)

plot subset of forcing data for 2002-2003 water year
prd = "2002-10-01/2003-09-30"
rvn_forcings_plot(fdata,prd)

}
\seealso{
\code{\link{rvn_forcings_read}} for the function used to read in the
forcings function data

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{diagnostics}
\keyword{forcing}
\keyword{plot}
