% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_get_prd}
\alias{rvn_get_prd}
\title{Determine period of data

rvn_get_prd either obtains the full period of a given xts object,
or checks the provided prd against an xts object for consistency and issues
a warning of an issue is found.

XXX to update error checking}
\usage{
rvn_get_prd(x, prd = NULL)
}
\arguments{
\item{x}{xts object}

\item{prd}{period argument in format YYYY-MM-DD/YYYY-MM-DD as a character}
}
\value{
{prd argument with warnings provided if needed}
}
\description{
Determine period of data

rvn_get_prd either obtains the full period of a given xts object,
or checks the provided prd against an xts object for consistency and issues
a warning of an issue is found.

XXX to update error checking
}
\examples{
data(rvn_hydrograph_data)

# get full valid prd argument for xts object
rvn_get_prd(rvn_hydrograph_data$hyd$Sub43_obs)

# check prd argument against xts object
rvn_get_prd(rvn_hydrograph_data$hyd$Sub43_obs, "2020-01-01/2020-02-01")
rvn_get_prd(rvn_hydrograph_data$hyd$Sub43_obs, "2002-24-01/2020-02-01")
rvn_get_prd(rvn_hydrograph_data$hyd$Sub43_obs, "20-24-01/2020-02-01")


}
\seealso{
\code{\link{rvn_theme_RavenR}} provides a theme for the RavenR package

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{get}
\keyword{prd}
\keyword{xts}
