% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_gridweights_read.R
\name{rvn_gridweights_read}
\alias{rvn_gridweights_read}
\title{Read in Raven GridWeights file}
\usage{
rvn_gridweights_read(ff = NA)
}
\arguments{
\item{ff}{full file path to the grid weights file}
}
\value{
\item{gridweights}{list with the number of HRUs, number of grid cells, and gridweights data frame}
}
\description{
rvn_gridweights_read is used to read in a Raven grid weights file.
}
\details{
This function expects a full file path to the grid weights file, then
reads in the file. The output is provided as a list containing the number of HRUs,
number of grid cells, and the grid weights data frame.
}
\examples{

# read in hydrograph sample csv data from RavenR package
ff <- system.file("extdata","run1_Hydrographs.csv", package="RavenR")

# read in Raven Hydrographs file, store into myhyd
myhyd <- rvn_hyd_read(ff)

# view contents
head(myhyd$hyd)
myhyd$units

}
\seealso{
\code{\link{rvn_gen_gridweights}} for generating a grid weights file from an HRU shapefile
and netcdf file.
}
