% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_hyd_extract.R
\name{rvn_hyd_extract}
\alias{rvn_hyd_extract}
\title{Extract function for Raven Hydrograph object}
\usage{
rvn_hyd_extract(subs = NA, hyd = NA, prd = NULL)
}
\arguments{
\item{subs}{column name for plotting/extracting}

\item{hyd}{full hydrograph data frame (including units) produced by hyd.read}

\item{prd}{time period for plotting, as string. See details}
}
\value{
returns a list with sim, obs, and inflow time series
 \item{sim}{model simulation for specified column and period}
 \item{obs}{observed data for specified column and period}
 \item{inflow}{inflow simulation for specified column and period}
}
\description{
rvn_hyd_extract is used for extracting data from the Raven hydrograph object.
Works for objects passed from rvn_hyd_read function (which reads the Hydrographs.csv file produced by the modelling framework Raven).
}
\details{
rvn_hyd_extract is used to extract the modelled and observed data from a Raven
hydrograph object by name reference. It is also easy to create plots of
modelled and observed data using this function. The simulated and observed
files are outputted regardless of whether a plot is created, for the
specified period.

The subs input is the name of the column desired for use; the most common
use of this will be for subbasin outflows, where the names will be of the
form "subXX", for example "sub24".

The hyd object is the full hydrograph object (hyd and units in one data
frame) created by the hyd.read function. Both the hyd and units are
required, since the units are placed onto the plots if one is created. This
is useful to at least see the units of the plotted variable, even if the
plot is later modified.

The prd input is used to specify a period for the plot and/or the data
output. The period should be specified as a string start and end date, of
the format "YYYY-MM-DD/YYYY-MM-DD", for example, "2006-10-01/2010-10-01". If
no period is supplied, the entire time series will be used.
}
\examples{
# read in hydrograph sample csv data from RavenR package
ff <- system.file("extdata","run1_Hydrographs.csv", package="RavenR")

# read in Raven Hydrographs file, store into myhyd
myhyd <- rvn_hyd_read(ff)

# no plot or observed data, specified period
flow_36 <- rvn_hyd_extract(subs="Sub36",myhyd)

attributes(flow_36)

# extract simulated and observed flows
sim <- flow_36$sim
obs <- flow_36$obs

# extract precipitation forcings
myprecip <- rvn_hyd_extract(subs="precip",hyd=myhyd)
myprecip <- myprecip$sim

# plot all components using rvn_hyd_plot
rvn_hyd_plot(sim,obs,precip=myprecip)

}
\seealso{
\code{\link{rvn_hyd_read}} for reading in the Hydrographs.csv file and
creating the object required in this function.
\code{\link{rvn_hyd_plot}} for conveniently plotting the output object contents onto the same figure.
}
