% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_hyd_plot.R
\name{rvn_hyd_plot}
\alias{rvn_hyd_plot}
\title{Plot Hydrograph}
\usage{
rvn_hyd_plot(
  sim = NULL,
  obs = NULL,
  inflow = NULL,
  precip = NULL,
  prd = NULL,
  winter_shading = T
)
}
\arguments{
\item{sim}{time series object of simulated flows}

\item{obs}{time series object of observed flows}

\item{inflow}{time series object of inflows to subbasin}

\item{precip}{time series object of precipitation}

\item{prd}{period to use in plotting}

\item{winter_shading}{optionally adds shading for winter months (default
TRUE)}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
rvn_hyd_plot creates a hydrograph plot for the supplied flow series, or
equivalently a stage plot for reservoir stages.
}
\details{
This function creates a hydrograph plot using the supplied time series; any
series not supplied will not be plotted. If the precip time series is
supplied, the secondary y axis will be used to plot the precip time series.

The function assumes that the supplied time series have the same length and
duration in time. If this is not true, then the defined period or period
calculated from the first available flow series will be used to determine
the plotting limits in time. If the data is used directly from Raven output,
this is not a concern. The supplied time series should be in xts format,
which again can be obtained directly by using the hyd.extract function.

The winter_shading argument will add a transparent cyan shading for the
December 1st to March 31st period in each year that is plotted.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# load sample hydrograph data, two years worth of sim/obs
data(rvn_hydrograph_data)
sim <- rvn_hydrograph_data$hyd$Sub36
obs <- rvn_hydrograph_data$hyd$Sub36_obs
precip <- rvn_hydrograph_data$hyd$precip

# create a nice hydrograph
rvn_hyd_plot(sim,obs,zero_axis=F)

# create a hydrograph with precip as well;
## range.mult=1.5 by default, leaves some overlap in plot axes
rvn_hyd_plot(sim,obs,precip=precip)

# create a hydrograph with precip as well for a specific subperiod
prd <- "2003-10-01/2004-10-01"
rvn_hyd_plot(sim,obs,precip=precip,prd=prd)

}
\seealso{
\code{\link{flow.spaghetti}} to create a spaghetti plot of annual
flow series

\code{\link{hyd.extract}} to extract time series from Raven objects

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{flow}
\keyword{hydrograph}
