% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_iscolour}
\alias{rvn_iscolour}
\title{Check Validity of Colour Representation}
\usage{
rvn_iscolour(x)
}
\arguments{
\item{x}{string or string vector of colour representations to test}
}
\value{
\item{y}{vector of TRUE or FALSE indicating whether the colour is valid}
}
\description{
rvn_iscolour checks whether a string or string vector contains valid colour representations
(in text or hexadecimal form). Useful in error checking colour arguments for functions.
}
\examples{

rvn_iscolour(c(NA, "black", "blackk", "1", "#00", "#000000"))
#   <NA>   black  blackk       1     #00 #000000
#   TRUE    TRUE   FALSE    TRUE   FALSE    TRUE

}
\seealso{
See original code on post in Stack Overflow
\href{https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation?utm_medium=organic&utm_source=google_rich_qa&utm_campaign=google_rich_qa}{
Check if character string is a valid color representation}

\code{\link{rvn_col_transparent}} for creating transparent colour codes
}
