% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_month_names}
\alias{rvn_month_names}
\title{Months in the Year vector}
\usage{
rvn_month_names(short = TRUE)
}
\arguments{
\item{short}{boolean to return shortened form of months}
}
\value{
{character array of month names}
}
\description{
rvn_month_names is used to return a character vector of months in the year
}
\examples{
rvn_month_names()
rvn_month_names(FALSE)

}
\seealso{
\code{\link{rvn_num_days}} for calculating the number of days in a
month
}
