% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_mos_names.R
\name{rvn_mos_names}
\alias{rvn_mos_names}
\title{Months in the Year vector}
\usage{
rvn_mos_names(short = T)
}
\arguments{
\item{short}{boolean to return shortened form of months}
}
\value{
\item{int}{number of days between the two days}
}
\description{
rvn_mos_names is used to return a string vector of months in the year
}
\examples{

months_of_the_year <- rvn_mos_names
months_of_the_year

rvn_mos_names(F)

}
\seealso{
\code{\link{num.days.month}} for calculating the number of days in a
month
}
\keyword{months}
\keyword{year}
