% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_netcdf_to_gridshp.R
\name{rvn_netcdf_to_gridshp}
\alias{rvn_netcdf_to_gridshp}
\title{Generate grid overlay from netCDF file

Takes the latitude-longitude cell coordinates from a netCDF file (assumed to be named 'lat' and 'long')
generates an estimate of the grid polygons associated with each netCDF cell and exports this to a shapefile (outshp)

Note that the function can fail due to bad netCDF file or inappropriate UTM zone;}
\usage{
rvn_netcdf_to_gridshp(ncfile, UTMzone, outshp)
}
\arguments{
\item{ncfile}{netCDF file with latitude and longitude variables}

\item{UTMzone}{UTM zone for exported shapefile (integer)}

\item{outshp}{name of output shapefile prefix (i.e., no .shp extension)}
}
\value{
\item{TRUE}{returns TRUE if executed properly. Also generates shapefile in folder}
}
\description{
Generate grid overlay from netCDF file

Takes the latitude-longitude cell coordinates from a netCDF file (assumed to be named 'lat' and 'long')
generates an estimate of the grid polygons associated with each netCDF cell and exports this to a shapefile (outshp)

Note that the function can fail due to bad netCDF file or inappropriate UTM zone;
}
\examples{

\dontrun{
ncfile <- system.file("extdata/GlenAllan.nc", package="RavenR")
UTMzone <- 17
outshp <- "maps/RDPS_Grid"
rvn_netcdf_to_gridshp(ncfile, UTMzone, outshp)
}


}
\seealso{
\code{\link{rvn_gen_gridweights}} for generating a shapefile of gridweights

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\author{
James R. Craig, University of Waterloo, 2019
}
\keyword{conversion}
\keyword{grid}
\keyword{netcdf}
\keyword{shapefile}
