% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_obsfile_create.R
\name{rvn_obsfile_rvt}
\alias{rvn_obsfile_rvt}
\title{Create Raven observation data (rvt) file}
\usage{
rvn_obsfile_rvt(filename, ts, SBID, typestr = "HYDROGRAPH", units = "m3/s")
}
\arguments{
\item{filename}{observation data file to be created, with .rvt extension}

\item{ts}{xts time series with single data column}

\item{SBID}{Subbasin ID for hydrographs and reservoir stages or HRU ID for observations of state variables (e.g., snow depth)}

\item{typestr}{Raven-recognized data type string: 'HYDROGRAPH', 'RESERVOIR_STAGE', 'RESERVOIR_INFLOW', 'RESERVOIR_NET_INFLOW',
or a state variable name (e.g., SOIL[0] or SNOW)}

\item{units}{units of the data (should be consistent with Raven units; neither Raven nor this routine checks or corrects}
}
\value{
nothing; produces file only
}
\description{
Creates an observation data file named filename from a continuous (gap-free) xts time series ts
}
\examples{
  mydata<-read.csv("JohnCreek.csv")
  flows<-timeseries.infill(mydata)
  rvn_obsfile_rvt("JohnCreek.rvt", flows, 13,typestr="HYDROGRAPH")

  # just weight March-October flows :
  wts<-obsweights.gen(flows,criterion="BETWEEN_CYCLIC",startdate="2000-03-01",enddate="2000-11-01")

  # and only after March 2003:
  wts<-obsweights.gen(flows,criterion="AFTER",startdate="2003-03-01")
  wts2=wts2*wts # product merges weights

  obsweightsfile.create("JohnCreekWts.rvt",wts2,13,typestr="HYDROGRAPH")

}
\seealso{
\code{\link{timeseries.infill}} \code{\link{obsweightsfile.create}}
See also the \href{http://raven.uwaterloo.ca/}{Raven website}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Raven}
\keyword{observations}
