% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_Ost_bestparams.R
\name{rvn_ost_bestparams}
\alias{rvn_ost_bestparams}
\title{Obtain set of best params from Ostrich data}
\usage{
rvn_ost_bestparams(
  ost_data = NA,
  best_num = 0.1,
  num_metrics = NA,
  remove_run = T
)
}
\arguments{
\item{ost_data}{data object from rvn_ost_read function (optional)}

\item{best_num}{indicates the method for selecting top best runs}

\item{num_metrics}{number of fit metrics used in Ostrich; will remove the
metrics if supplied}

\item{remove_run}{boolean for removing the 'Run' column in returned data}
}
\value{
\item{ost_bestruns}{Ostrich trials with the top best objective
function}
}
\description{
rvn_ost_bestparams is used to extract the best trials from the Ostrich data,
based on the objective function value.
}
\details{
rvn_ost_bestparams extracts the runs with the best objective function value from
the data supplied. This function may be useful in comparing the parameters
of an optimization trial to check for equifinality in models.

ost_data is the data object created from rvn_ost_read. If not supplied, this
function will look for the OstModel0.txt file in the current directory to
read data from.

best_num indicates the method used to get the top best trials. If best_num
is a double less than 1, this function will return the top percentage of
best trials. If best_num is an integer, the function will return the top X
trials in the file, where X=best_num.

num_metrics refers to the number of fit metrics used in the Ostrich
calibration, of which there should be at least one. If this value is
supplied, the function will remove the metric columns from the returned
data.

remove_run is a boolean to indicate whether the run column should be
removed.
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{

# locate sample OstModel0.txt file in RavenR package
ff <- system.file("extdata","OstModel0.txt", package="RavenR")

# read in file
mydata <- rvn_ost_read(ff)

# top 30 percent of runs, no removal of metrics
mybest <- rvn_ost_bestparams(mydata,best_num=0.3)
head(mybest)

# minimalist function run (requires OstModel0.txt in current directory)
mybest <- rvn_ost_bestparams(mydata)
head(mybest)
}

}
\seealso{
\code{\link{rvn_ost_read}} for reading in Ostrich files

\code{\link{rvn_ost_plot}} for creating various plots with Ostrich data

\code{\link{ost_params_compare}} for comparing parameters between multiple
Ostrich calibration jobs

Download Ostrich at
\href{http://www.eng.buffalo.edu/~lsmatott/Ostrich/OstrichMain.html}{Shawn
Matott's web page}

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for other software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Ostrich}
\keyword{best}
\keyword{calibration}
\keyword{parameter}
