% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_res_dygraph.R
\name{rvn_res_dygraph}
\alias{rvn_res_dygraph}
\title{Plot Raven reservoir/lake stage time series using dygraph}
\usage{
rvn_res_dygraph(resdata, timezone = "UTC", basins = "", figheight = 400)
}
\arguments{
\item{resdata}{reservoir stage time series data structure generated by res.read() routine}

\item{timezone}{data timezone; defaults to UTC}

\item{basins}{list of subbasin names from reservoir file. Each subbasin creates separate dygraph plots}

\item{figheight}{height of figure, in pixels}
}
\value{
a list of plot handles to dygraph plots

See also \href{http://raven.uwaterloo.ca/}{Raven page}
}
\description{
rvn_res_dygraph plots modeled vs observed stage plots when supplied with reservoir stage data
structure read using hyd.read()
}
\examples{

# read in Raven Reservoir Stages file
resdata<-res.read("C:/TEMP/ReservoirStages.csv")

# view contents for all subbasins as dyGraph
res<-res.dyGraphs(resdata,basins="sub11")
res[0]
res[1]

}
\keyword{Raven}
\keyword{plotting}
\keyword{reservoir}
\keyword{stage}
