% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_res_init.R
\name{rvn_res_init}
\alias{rvn_res_init}
\title{Create initial conditions file for Reservoirs}
\usage{
rvn_res_init(ff = NA, initial_stage = 0)
}
\arguments{
\item{ff}{full file path to the reservoir information file}

\item{initial_stage}{an optional double for percentage of maximum stage to
use as initial condition; default 0.0}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
rvn_res_init is used to write an initial conditions (rvc) format file for Raven,
with the calculated reservoir stages written in.
}
\details{
This function is used to write an initial conditions format file for Raven
with the relevant initial reservoir stages. This file can be used directly
as the model rvc file, or one may copy and paste the information into a
separate rvc file for use (i.e. if there is other information to be included
in the model rvc file).

The supplied file in ff should be a csv file with the format headers of:
SUBID,STAGE,FLOW,VOL,AREA. The header text is not important as the function
assumes the order only, however some header must exist. The SUBID will be
the integer of the subbasin where the reservoir is located, and will repeat
the value for each point in the stage relationship provided. This is the
only required variable. This format is the same as is requried for the
res.fit function (thus the extra columns that are not used by this
function).

Each point in the stage relationship table is used, so if there exist
multiple dry inflow points, the minimum stage of these points will be used
as the 'minimum' stage.

Also note that the initial stage must be between 0 and 1, so that the
initial stage is not less than the dry stage or greater than the maximum.
}
\examples{
# warning: example not run, sample example for associated files only
\dontrun{
# set working directory to file location
dir = "C:/temp/test"
ff <- paste0(dir,"/","reservoir data.csv")

rvn_res_init(ff,initial_stage=0.4)
}

}
\seealso{
\code{\link{rvn_res_read}} for reading in the ReservoirStages.csv file

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{conditions}
\keyword{initial}
\keyword{relationship}
\keyword{reservoir}
\keyword{stage}
