% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_res_plot.R
\name{rvn_res_plot}
\alias{rvn_res_plot}
\title{Plot Reservoir Stage}
\usage{
rvn_res_plot(
  sim = NULL,
  obs = NULL,
  inflow = NULL,
  precip = NULL,
  prd = NULL,
  winter_shading = T
)
}
\arguments{
\item{sim}{time series object of simulated stage}

\item{obs}{time series object of observed stage}

\item{inflow}{time series object of inflow to subbasin}

\item{precip}{time series object of precipitation}

\item{prd}{period to use in plotting}

\item{winter_shading}{optionally adds shading for winter months (default
TRUE)}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
rvn_res_plot creates a reservoir stage plot for the supplied flow series, or
equivalently a stage plot for reservoir stages.
}
\details{
This function creates a reservoir stage plot using the supplied time series;
any series not supplied will not be plotted. If the precip time series is
supplied, the secondary y axis will be used to plot the precip time series.

The function assumes that the supplied time series have the same length and
duration in time. If this is not true, then the defined period or period
calculated from the first available flow series will be used to determine
the plotting limits in time. If the data is used directly from Raven output,
this is not a concern. The supplied time series should be in xts format,
which again can be obtained directly by using the rvn_res_extract function.

The winter_shading argument will add a transparent cyan shading for the
December 1st to March 31st period in each year that is plotted.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{
# create a nice reservoir stage plot
rvn_res_plot(sim,obs,zero_axis=F)

# create a reservoir stage plot with precip as well
rvn_res_plot(sim,obs,precip=precip)

# create a reservoir stage plot with precip as well for a specific subperiod
prd <- "2003-10-01/2005-10-01"
rvn_res_plot(sim,obs,precip=precip,prd=prd)

}
\seealso{
\code{\link{rvn_hyd_read}} for reading in the Hydrographs.csv file

\code{\link{rvn_res_extract}} to extract time series from Raven objects

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{reservoir}
\keyword{stage}
