% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvh_HRUfromSHP.R
\name{rvn_rvh_HRUfromSHP}
\alias{rvn_rvh_HRUfromSHP}
\title{Generate Raven HRU Data.frame from Shapefile}
\usage{
rvn_rvh_HRUfromSHP(
  hrushp,
  HRUID_col = NA,
  area_col = NA,
  elev_col = "Elevation",
  lat_col = "Latitude",
  long_col = "Longitude",
  SBID_col = "SBID",
  landclass_col = "LandUse",
  vegclass_col = "Vegetation",
  soilclass_col = "SoilProfile",
  terrain_col = NA,
  aquifer_col = NA,
  slope_col = "Slope",
  aspect_col = "Aspect"
)
}
\arguments{
\item{hrushp}{sf object or data.frame with HRU columns. Must have geometry if area_col is NA.}

\item{HRUID_col}{character, HRU ID column name (optional)}

\item{area_col}{character, HRU area column name (optional)}

\item{elev_col}{character, HRU elevation column name}

\item{lat_col}{character, HRU latitude column name}

\item{long_col}{character, HRU longitude column name}

\item{SBID_col}{character, HRU membership subbasin id column name}

\item{landclass_col}{character, HRU land class column name}

\item{vegclass_col}{character, HRU vegetation class column name}

\item{soilclass_col}{character, HRU soil type column name}

\item{terrain_col}{character, HRU terrain type column name (optional)}

\item{aquifer_col}{character, HRU aquifer class column name (optional)}

\item{slope_col}{character, HRU slope column name}

\item{aspect_col}{character, HRU aspect column name}
}
\value{
data.frame of HRU properties for writing to RVH file
}
\description{
For creating a data.frame of Raven HRU properties from a shapefile with the appropriate columns.
Compatible with \code{\link{rvn_rvh_write}}. Generic column names are assumed for shapefile, but can
also be passed as arguments.
}
