% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvh_SBfromSHP.R
\name{rvn_rvh_SBfromSHP}
\alias{rvn_rvh_SBfromSHP}
\title{Generate Raven SubBasin Data.frame from Shapefile}
\usage{
rvn_rvh_SBfromSHP(
  sbshp,
  SBID_col = NA,
  name_col = NA,
  downstream_col = "Downstream_ID",
  profile_col = "Profile",
  length_col = "ReachLength",
  gauged_col = "Gauged"
)
}
\arguments{
\item{sbshp}{sf object or data.frame with subbasin columns}

\item{SBID_col}{character, subbasin ID column name (optional)}

\item{name_col}{character, subbasin name column name (optional)}

\item{downstream_col}{character, subbasin downstream subbasin column name}

\item{profile_col}{character, subbaisn reach profile column name}

\item{length_col}{character, subbasin reach length column name}

\item{gauged_col}{character, subbasin boolean gauged flag column name}
}
\value{
data.frame of subbasin properties for writing to RVH file
}
\description{
For creating a data.frame of Raven subbasin properties from a shapefile with the appropriate columns.
Compatible with \code{\link{rvn_rvh_write}}. Generic column names are assumed for shapefile, but can
also be passed as arguments.
}
