% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvi_connections.R
\name{rvn_rvi_connections}
\alias{rvn_rvi_connections}
\title{Generate Hydrological process connections list}
\usage{
rvn_rvi_connections(
  rvi,
  ProcConDataFile = system.file("extdata", "RavenProcessConnections.dat", package =
    "RavenR")
)
}
\arguments{
\item{rvi}{data object generated from the \code{\link{rvn_rvi_read}} routine}

\item{ProcConDataFile}{(optional) path to RavenProcesConnections.dat file}
}
\value{
Returns a list with two items:
\item{connections}{a a dataframe of all of the process connections Includes the following data columns:
process type, algorithm, 'from' compartment, 'to' compartment, and conditional}
\item{AliasTable}{a table of aliases (unchanged from the supplied \code{rvi$AliasTable})}
}
\description{
This routine reads in a hydrologic process list from \code{\link{rvn_rvi_read}} and generates
the list of hydrologic process connections.
}
\details{
Relies on a valid and up-to-date RavenProcessConnections.dat file.
This file is provided with the RavenR package, but
may be overridden by a more recent file if provided manually.
}
\examples{
rvi <- rvn_rvi_read(system.file("extdata","Nith.rvi", package="RavenR"))

rvi_conn <- rvn_rvi_connections(rvi)
head(rvi_conn$connections)
head(rvi_conn$AliasTable)

}
\seealso{
\code{\link{rvn_rvi_read}} to read a .rvi file and generate an rvi object, and
\code{\link{rvn_rvi_process_ggplot}} or \code{\link{rvn_rvi_process_diagrammer}} to plot the process network produced in this function.

See also the \href{https://raven.uwaterloo.ca/}{Raven page}
}
\author{
James R. Craig, University of Waterloo
}
