% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvi_hpnet_plot.R
\name{rvn_rvi_hpnet_plot}
\alias{rvn_rvi_hpnet_plot}
\title{Plot Raven hydrologic process network}
\usage{
rvn_rvi_hpnet_plot(connections, pdfout = "network.pdf")
}
\arguments{
\item{connections}{a dataframe of from-to connections generated using rvi.connections()}

\item{pdfout}{name of pdf file to save the network plot to}
}
\value{
{TRUE}{returns TRUE if executed properly. Also generates a pdf plot in working directory.}
}
\description{
This routine takes a connections datafrom generated using rvi.connections()
and plots the connections information as a network graph.
}
\note{
tries to follow basic network structure, accomodates unrecognized state variables on LHS of plot
}
\examples{
 \dontrun{
  # sample workflow of rvh.read

  rvi<-rvi.read("example.rvi")

  # get number of Hydrologic processes
  numProcss<-nrow(rvi$HydProcTable)
}
}
\seealso{
rvi.connections rvi.read
See also the \href{http://raven.uwaterloo.ca/}{Raven page}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Hydrologic}
\keyword{Processes}
\keyword{Raven}
\keyword{rvi}
