% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvi_process_diagrammer.R
\name{rvn_rvi_process_diagrammer}
\alias{rvn_rvi_process_diagrammer}
\title{Plot Raven hydrologic process network using DiagrammeR}
\usage{
rvn_rvi_process_diagrammer(connections, pdfout = NULL)
}
\arguments{
\item{connections}{a dataframe of from-to connections generated using rvn_rvi_connections()}

\item{pdfout}{name of pdf file to save the network plot to, if null no PDF is generated}
}
\value{
\code{d1} returns DiagrammeR object. Also generates a .pdf file in working directory if pdfplot argument is not NULL.
}
\description{
This routine takes a connections data from generated using rvn_rvi_connections()
and returns the connections information as a DiagrammeR object.
}
\details{
Uses the output from the \code{\link{rvn_rvi_connections}} function to generate the plot.

Note that the output can be also be plotted using the \code{\link{render_graph}} function
in the DiagrammeR library.

The outputted DiagrammeR object may also have aesthetics modified with various commands from the
same library, if desired.
}
\note{
tries to follow basic network structure, accommodates unrecognized state variables on LHS of plot
}
\examples{
rvi <- rvn_rvi_read(system.file("extdata","Nith.rvi", package="RavenR"))
conn <- rvn_rvi_connections(rvi)

rvn_rvi_process_diagrammer(conn) \%>\%
render_graph()

}
\seealso{
\code{\link{rvn_rvi_connections}} to generate connections table from an rvi object

See also the \href{http://raven.uwaterloo.ca/}{Raven page}
}
