% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvi_process_plot.R
\name{rvn_rvi_process_plot}
\alias{rvn_rvi_process_plot}
\title{Plot Raven hydrologic process network}
\usage{
rvn_rvi_process_plot(connections, plot = TRUE, pdfout = NULL)
}
\arguments{
\item{connections}{a dataframe of from-to connections generated using rvn_rvi_connections()}

\item{plot}{boolean whether to print the plot, default TRUE}

\item{pdfout}{name of pdf file to save the network plot to, if null no PDF is generated#'}
}
\value{
{p1}{returns ggplot object. Also generates a .pdf file in working directory if pdfplot argument is not NULL.}
}
\description{
This routine takes a connections datafrom generated using rvn_rvi_connections()
and plots the connections information as a network graph.
}
\note{
tries to follow basic network structure, accomodates unrecognized state variables on LHS of plot
}
\examples{
 rvi <- rvn_rvi_read(system.file("extdata","Nith.rvi", package="RavenR"))

 # get number of Hydrologic processes
 nrow(rvi$HydProcTable)

 conn <- rvn_rvi_connections(rvi)
 rvn_rvi_process_plot(conn)

}
\seealso{
\code{\link{rvn_rvi_connections}} to generate connections table from an rvi object

See also the \href{http://raven.uwaterloo.ca/}{Raven page}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Hydrologic}
\keyword{Processes}
\keyword{Raven}
\keyword{rvi}
