% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvi_process_visplot.R
\name{rvn_rvi_process_visplot}
\alias{rvn_rvi_process_visplot}
\title{Plot Raven hydrologic process network using visNetwork}
\usage{
rvn_rvi_process_visplot(rvi_conn, custom_label = NULL, default_label = FALSE)
}
\arguments{
\item{rvi_conn}{a list of connections and AliasTable, provided by \code{rvn_rvi_connections}}

\item{custom_label}{(optional) a two-columns matrix/data.frame in which the first and the second columns are  equal the hydrologic compartment labels in the \code{rvi_conn} and their corresponding replacement labels respectively provided that \code{default_label = FALSE}}

\item{default_label}{(optional) logical. if \code{TRUE} an internal default labels are used as the compartments names given that \code{custom_label = NULL}}
}
\value{
{p1}{returns visNetwork plot}
}
\description{
This routine takes a connections data from generated using \code{rvn_rvi_connections}
and returns the connections information as a network graph.
}
\examples{

## library(GGally)
## to adjust handling of ProcCondDataFile in function

  rvn_rvi_read(system.file("extdata","Nith.rvi", package="RavenR")) \%>\%
  rvn_rvi_connections() \%>\%
  rvn_rvi_process_visplot(.,default_label=TRUE)

}
\seealso{
\code{\link{rvn_rvi_connections}} to generate connections table from an rvi object

See also the \href{http://raven.uwaterloo.ca/}{Raven page}
}
