% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvt_ECmet_fromStns.R
\name{rvn_rvt_ECmet_fromStns}
\alias{rvn_rvt_ECmet_fromStns}
\title{Generate Raven .rvt files from EC Station data}
\usage{
rvn_rvt_ECmet_fromStns(station_tbl, start, end, interval, forcing_set = 1, ...)
}
\arguments{
\item{station_tbl}{tibble of Station info from weathercan::stations_search(). Can also be data.frame with station_ids column.}

\item{start}{Date/Character. The start date of the data in YYYY-MM-DD format (applies to all stations_ids).}

\item{end}{Date/Character. The end date of the data in YYYY-MM-DD format (applies to all station_ids).}

\item{interval}{Character. Interval of the data, one of "hour", "day", "month".}

\item{forcing_set}{Integer. Specifies the set of forcings to print to file, see \link{rvn_rvt_ECmet} (optional)}

\item{...}{Additional arguments passed to \link{rvn_rvt_ECmet}}
}
\value{
TRUE returns TRUE if function executed successfully.
}
\description{
Uses the weathercan package to download data and write the .rvt file using teh data.
}
\examples{
# Built for pipes
weathercan::stations_search("Waterloo", interval="day") \%>\%
            filter(station_id \%in\% c(4832, 4831)) \%>\%
            rvn_rvt_ECmet_fromStns(start="1990-10-01", end="2000-09-30", interval="day")
}
\author{
Leland Scantlebury
}
