% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_rvt_mappings}
\alias{rvn_rvt_mappings}
\title{Provide mappings for rvt files}
\usage{
rvn_rvt_mappings()
}
\value{
\code{TRUE} if executed successfully.
}
\description{
Provides global variables to the environment regarding the mapping of rvt_type and data_type
that are used in the \code{\link{rvn_rvt_write}} function.
}
\examples{
rvn_rvt_mappings()

# view global variables
rvt_mapping
rvt_data_type_mapping

}
\seealso{
\code{\link{rvn_rvt_write}}
}
