% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvt_obsfile.R
\name{rvn_rvt_obsfile}
\alias{rvn_rvt_obsfile}
\title{Create Raven observation data (rvt) file}
\usage{
rvn_rvt_obsfile(filename, ts, SBID, typestr = "HYDROGRAPH", units = "m3/s")
}
\arguments{
\item{filename}{observation data file to be created, with .rvt extension}

\item{ts}{xts time series with single data column}

\item{SBID}{Subbasin ID for hydrographs and reservoir stages or HRU ID for observations of state variables (e.g., snow depth)}

\item{typestr}{Raven-recognized data type string: 'HYDROGRAPH', 'RESERVOIR_STAGE', 'RESERVOIR_INFLOW', 'RESERVOIR_NET_INFLOW',
or a state variable name (e.g., SOIL[0] or SNOW)}

\item{units}{units of the data (should be consistent with Raven units; neither Raven nor this routine checks or corrects}
}
\value{
{TRUE}{returns TRUE if function runs properly}
}
\description{
Creates an observation data file named filename from a continuous (gap-free) xts time series ts
}
\examples{
\dontrun{

  ## this example needs to be totally cleaned up!!!

  mydata<-read.csv("JohnCreek.csv")
  flows<-timeseries.infill(mydata)
  rvn_obsfile_rvt("JohnCreek.rvt", flows, 13,typestr="HYDROGRAPH")

  # just weight March-October flows :
  wts<-obsweights.gen(flows,criterion="BETWEEN_CYCLIC",startdate="2000-03-01",enddate="2000-11-01")

  # and only after March 2003:
  wts<-obsweights.gen(flows,criterion="AFTER",startdate="2003-03-01")
  wts2=wts2*wts # product merges weights

  rvn_rvt_obsfile("JohnCreekWts.rvt",wts2,13,typestr="HYDROGRAPH")
  }

}
\seealso{
\code{\link{rvn_ts_infill}} for infilling time series, and  \code{\link{rvn_rvt_obsweights}} to write an rvt observation weights file
See also the \href{http://raven.uwaterloo.ca/}{Raven website}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Raven}
\keyword{observations}
