% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvt_obsweights.R
\name{rvn_rvt_obsweights}
\alias{rvn_rvt_obsweights}
\title{Create Raven observation data weight (rvt) file}
\usage{
rvn_rvt_obsweights(filename, wts, SBID, typestr = "HYDROGRAPH")
}
\arguments{
\item{filename}{observation data file to be created, with .rvt extension}

\item{wts}{xts time series with single data column of observation weights}

\item{SBID}{Subbasin ID for hydrographs and reservoir stages or HRU ID for observations of state variables (e.g., snow depth)}

\item{typestr}{Raven-recognized data type string: 'HYDROGRAPH', 'RESERVOIR_STAGE', 'RESERVOIR_INFLOW', 'RESERVOIR_NET_INFLOW',
or a state variable name (e.g., SOIL[0] or SNOW)}
}
\value{
{TRUE}{returns TRUE if function runs properly}
}
\description{
Creates an observation data weights file named filename from a continuous (gap-free) xts time series ts
with values between 0 and 1
}
\examples{
 # locate hydrograph sample csv data from RavenR package
 ff <- system.file("extdata","run1_Hydrographs.csv", package="RavenR")

 # read in Raven Hydrographs file, store into mydata
 mydata <- rvn_hyd_read(ff)

 # generate rvt file using just observations from Subbasin ID 36
 flows <- rvn_ts_infill(mydata$hyd$Sub36_obs)
 rvn_rvt_obsfile("run1_Hydrographs.rvt", flows, 36, typestr = "HYDROGRAPH")

 # weight March-October flows:
 wts <- rvn_gen_obsweights(flows,criterion = "BETWEEN_CYCLIC", startdate="2000-03-01", enddate="2003-11-01")

 # and only after March 2003:
 wts2 <- rvn_gen_obsweights(flows,criterion = "AFTER", startdate="2003-03-01")
 wts2 <- wts2*wts # product merges weights

 # write observation weights to rvt file
 rvn_rvt_obsweights("run1_Hydrographs_wts.rvt", wts2, 36, typestr="HYDROGRAPH")

}
\seealso{
\code{\link{rvn_obsfile_rvt}} \code{\link{obsweights.gen}}
See also the \href{http://raven.uwaterloo.ca/}{Raven website}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Raven}
\keyword{file}
\keyword{observations}
\keyword{rvt}
\keyword{weights}
\keyword{write}
