% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_stringpad}
\alias{rvn_stringpad}
\title{Pads string with spaces, either right or left justified}
\usage{
rvn_stringpad(string, width, just = "r")
}
\arguments{
\item{string}{Text string}

\item{width}{Number of characters total, including desired spaces}

\item{just}{'r' for right, 'l' for left}
}
\value{
{Padded string}
}
\description{
Pad string with spaces, justified on either the left or right
}
\examples{
# Returns "   To the right"
rvn_stringpad('To the right', 15, just='r')

# Returns "Padded    "
rvn_stringpad('Padded', 10, just='l')
}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
