% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_substrLeft}
\alias{rvn_substrLeft}
\title{substring from the Left}
\usage{
rvn_substrLeft(x, n)
}
\arguments{
\item{x}{a string to manipulate}

\item{n}{number of characters to remove from the left side of the string}
}
\description{
rvn_substrLeft returns n characters from the left side of the supplied string x.
}
\examples{

rvn_substrLeft("hello world",3)
# returns "hel"

}
\seealso{
\code{\link{rvn_substrRight}} for using n characters from right side of string

\code{\link{rvn_substrMRight}} for removing n characters from the right side of a string
}
