% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_substrMLeft}
\alias{rvn_substrMLeft}
\title{substring minus characters from the Left}
\usage{
rvn_substrMLeft(x, n)
}
\arguments{
\item{x}{a string to manipulate}

\item{n}{number of characters to remove from the left side of the string}
}
\description{
rvn_substrMLeft returns a string x with n characters removed from the left side
of the string.
}
\examples{

rvn_substrMLeft("hello world",3)
# returns "lo world"

}
\seealso{
\code{\link{rvn_substrRight}} for using n characters from the right
side of string,

\code{\link{rvn_substrLeft}} for using n characters from the left side of
string

\code{\link{rvn_substrMRight}} for removing n characters from the right side of
a string
}
