% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_substrRight}
\alias{rvn_substrRight}
\title{substring from the Right}
\usage{
rvn_substrRight(x, n)
}
\arguments{
\item{x}{a string to manipulate}

\item{n}{number of characters to use from the right side of the string}
}
\description{
rvn_substrRight returns n characters from the right side of the supplied string
x.
}
\examples{

rvn_substrRight("hello world",3)
# returns "rld"

}
\seealso{
\code{\link{rvn_substrLeft}} for using n characters from the left side
of string

\code{\link{rvn_substrMRight}} for removing n characters from the right side of
a string

\code{\link{rvn_substrMLeft}} for removing n characters from the left side of a
string
}
