% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_theme_RavenR.R
\name{rvn_theme_RavenR}
\alias{rvn_theme_RavenR}
\title{RavenR ggplot theme}
\usage{
rvn_theme_RavenR()
}
\value{
returns a theme for use in ggplot2 figures
}
\description{
rvn_theme_RavenR makes the general Raven R Theme for all ggplots
}
\details{
This function sets up the default theme for all ggplots generated using a
built in Raven R function. Made by adjusting the built in theme_bw().
}
\examples{
# generate a basic ggplot and apply the RavenR theme
library(ggplot2)
ggplot(data=cars, aes(x=speed, y=dist))+
geom_point()+
rvn_theme_RavenR()

}
\seealso{
\code{\link{rvn_annual_volume}} to create a scatterplot of annual flow
volumes.
}
