% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_write_Raven_label.R
\name{rvn_write_Raven_label}
\alias{rvn_write_Raven_label}
\title{Writes common Raven labeled line to file, with optional value (appends)}
\usage{
rvn_write_Raven_label(label, filename, value = NULL, digits = NULL)
}
\arguments{
\item{label}{character, (e.g. "SoilClasses")}

\item{filename}{character, file name/path to write to, with extension}

\item{value}{numeric or character, corresponding value written after label (optional)}

\item{digits}{Number of digits to round value to (optional)}
}
\value{
TRUE returns TRUE if executed successfully
}
\description{
Writes common Raven labeled line to file, with optional value (appends)
}
\examples{
rvn_write_Raven_label('Duration', 'Hogwarts.rvi', value=365)

# cleanup temporary file
unlink("Hogwarts.rvi")

}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
