% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_yearly_quantile_plot.R
\name{rvn_yearly_quantile}
\alias{rvn_yearly_quantile}
\title{Calculates Yearly Median, Upper and Lower Quantiles of Flow}
\usage{
rvn_yearly_quantile(
  hgdata,
  minyear = NULL,
  maxyear = NULL,
  Qlower = 0.1,
  Qupper = 0.9,
  water_year = T
)
}
\arguments{
\item{hgdata}{Time series object of observed or simulated flows}

\item{minyear}{Integer of minimum year (optional)}

\item{maxyear}{Integer of maximum year (optional)}

\item{Qlower}{Decimal percentage of lower quantile value (default 0.1)}

\item{Qupper}{Decimal percentage of upper quantile value (default 0.9)}
}
\value{
Time series object of monthly median and quantile values
}
\description{
Calculates Yearly Median, Upper and Lower Quantiles of Flow
}
\examples{
data("rvn_hydrograph_data")

# Pull out a specific hydrograph
hgdata <- hydrograph.data$hyd$Sub36

# Calculate quantiles
qdat <- rvn_yearly_quantile(hgdata)

}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
