% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMap.animate.R
\name{SBMap.animate}
\alias{SBMap.animate}
\title{Create Subbasin Map Animation}
\usage{
SBMap.animate(shpfilename, subIDcol, plot.daterange, cust.data,
  leg.title = "Legend", leg.pos = "bottomleft", normalize.data = TRUE,
  colour.scheme = "White-Blue", num.classes = 5, invalid.stop = TRUE,
  basins.label = "subID", plot.title = "",
  gif.filename = "subbasin_animated_plot.gif", transition.speed = 50)
}
\arguments{
\item{TBD}{params to be added}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
# description to be added
}
\details{
Creates animated subbasin plot from custom data and specified time period

Requires ImageMagick to be installed on your system

see informal parameter description here
}
\examples{

# Warning: example not run
# Example to be cleaned up

\dontrun{
plot.daterange <- '2000-01-01/2000-01-31'
cust.data <- custom.read('PRECIP_Daily_Average_BySubbasin.csv')
leg.title <- 'Legend: Precip (mm/d)'
plot.title <- 'Precipitation (mm/d)'

# create GIF of custom data plots
SBmap.plot.animate(shpfilename,subIDcol,plot.daterange,cust.data,plot.title=plot.title,leg.title=leg.title,gif.filename = 'Precip_mmd_Jan2000.gif')

transition.speed <- 30
plot.daterange <- '2000-02-01/2000-02-28'
cust.data <- custom.read('SNOW_Daily_Average_BySubbasin.csv')
plot.title <- 'Snow Cover (mm)'
leg.title <- 'Legend: Snow (mm)'
colour.scheme <- "Blue-White"
num.classes=7

# create GIF of custom data plots
SBmap.plot.animate(shpfilename,subIDcol,plot.daterange,cust.data,plot.title=plot.title,colour.scheme=colour.scheme,
                   num.classes=num.classes,leg.title=leg.title,transition.speed = transition.speed,gif.filename = 'Snow_mm_Jan2000.gif')
}

}
\seealso{
\code{\link{subbasinNetwork.plot}} to create network plots

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{map}
\keyword{plot}
\keyword{subbasin}
