% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMap.plot.R
\name{SBMap.plot}
\alias{SBMap.plot}
\title{Create Subbasin Map}
\usage{
SBMap.plot(shpfilename, subIDcol, plot.date, cust.data, leg.title = "Legend",
  leg.pos = "bottomleft", normalize.data = TRUE,
  colour.scheme = "White-Blue", num.classes = 5, invalid.stop = TRUE,
  basins.label = "subID", plot.title = "")
}
\arguments{
\item{TBD}{params to be added}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
# description to be added
}
\details{
Creates single static subbasin plot from custom data

see informal parameter description here

shpfilename - file path to shapefile.shp file
}
\examples{

# Warning: example not run
# Example to be cleaned up

\dontrun{
shpfilename <- 'shp/subbasins.shp'
subIDcol <- 'subID'
plot.date <- "2010-01-02"
cust.data <- custom.read('PRECIP_Daily_Average_BySubbasin.csv')
leg.title <- 'Legend - Precip (mm)'
leg.pos <- 'topright'
normalize.data <- TRUE
colour.scheme <- 'White-Blue'
num.classes <- 5
invalid.stop <- TRUE
basins.label <- 'subID'
# 'None', 'subID', 'value'
plot.title <- 'Precipitation (mm/d)'


# create singe plot
SBMap.plot(shpfilename,subIDcol,plot.date,cust.data,plot.title=plot.title)

cust.data <- custom.read('SNOW_Daily_Average_BySubbasin.csv')
plot.title <- 'Snow (mm)'

# create singe plot
SBMap.plot(shpfilename,subIDcol,plot.date,cust.data,plot.title=plot.title)
}

}
\seealso{
\code{\link{subbasinNetwork.plot}} to create network plots

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{map}
\keyword{plot}
\keyword{subbasin}
