% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry_client.R
\name{sentry_client}
\alias{sentry_client}
\title{Create Sentry Client}
\usage{
sentry_client(dsn, user = NULL, version = NULL)
}
\arguments{
\item{dsn}{set Data Source Name}

\item{user}{set user context}

\item{version}{protocol version}
}
\description{
\code{sentry_client} Creates a sentry client
}
