% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry_client.R
\name{set_tags}
\alias{set_tags}
\title{Set tags for sentry, will be added to future sentry logs}
\usage{
set_tags(...)
}
\arguments{
\item{tags}{named arguments with key, value combinations of tags. existing tags will be overwritten}
}
\description{
Set tags for sentry, will be added to future sentry logs
}
