% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop.R
\name{stop}
\alias{stop}
\title{Custom stop function to intercept errors}
\usage{
stop(..., call. = TRUE, domain = NULL)
}
\arguments{
\item{...}{other arguments}

\item{call.}{call}

\item{domain}{domain}
}
\description{
Custom stop function to intercept errors
}
