% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_file_write_utils.R
\name{stringPad}
\alias{stringPad}
\title{Pads string with spaces, either right or left justified}
\usage{
stringPad(string, width, just = "r")
}
\arguments{
\item{string}{Text string}

\item{width}{Number of characters total, including desired spaces}

\item{just}{'r' for right, 'l' for left}
}
\value{
Padded string
}
\description{
Pads string with spaces, either right or left justified
}
\examples{
# Returns '   To the right'
stringPad('To the right', 15, just='r')

}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
