% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subbasinNetwork.plot.R
\name{subbasinNetwork.plot}
\alias{subbasinNetwork.plot}
\title{Plot Raven subbasin network.}
\usage{
subbasinNetwork.plot(SBtable, labeled = FALSE)
}
\arguments{
\item{SBtable}{a valid table of Raven subbasins, obtained from \code{\link{rvh.read}}}

\item{labeled}{TRUE if the nodes are labeled with the SubBasin ID, SBID}
}
\value{
TRUE if completed
}
\description{
Takes the information gathered from an .rvh file via the function \code{\link{rvh.read}} and generates a plot of the subbasin network,
where nodes are located at SubBasin lat-long centroids, and edge widths of the network correspond to contributing upstream area.
}
\examples{
 # warning: example not run in compiling package
\dontrun{
 # read in rvh file
 rvh<-rvh.read("example.rvh")

 # create network plot of watershed structure from rvh file
 subbasinNetwork.plot(rvh$SBtable)
}

}
\seealso{
See also the \href{http://raven.uwaterloo.ca/}{Raven page}
\code{\link{rvh.read}} to read in the subbasin data table
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Network}
\keyword{Plot}
\keyword{Raven}
\keyword{Stream}
