% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrLeft.R
\name{substrLeft}
\alias{substrLeft}
\title{substring from the Left}
\usage{
substrLeft(x, n)
}
\arguments{
\item{x}{a string to manipulate}

\item{n}{number of characters to use from the left side of the string}
}
\description{
substrLeft returns n characters from the left side of the supplied string x.
}
\examples{

substrLeft("hello world",3)
# returns "hel"

}
\seealso{
\code{\link{substrRight}} for using n characters from right side of
string,

\code{\link{substrMRight}} for removing n characters from the right side of
a string,

\code{\link{substrMLeft}} for removing n characters from the left side of a
string
}
\keyword{left}
\keyword{string}
