% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrMLeft.R
\name{substrMLeft}
\alias{substrMLeft}
\title{substring minus characters from the Left}
\usage{
substrMLeft(x, n)
}
\arguments{
\item{x}{a string to manipulate}

\item{n}{number of characters to remove from the left side of the string}
}
\description{
substrMLeft returns a string x with n characters removed from the left side
of the string.
}
\examples{

substrMLeft("hello world",3)
# returns "lo world"

}
\seealso{
\code{\link{substrRight}} for using n characters from the right
side of string,

\code{\link{substrLeft}} for using n characters from the left side of
string,

\code{\link{substrMRight}} for removing n characters from the right side of
a string,
}
\keyword{left}
\keyword{minus}
\keyword{string}
