% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{substring from the Right}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{a string to manipulate}

\item{n}{number of characters to use from the right side of the string}
}
\description{
substrRight returns n characters from the right side of the supplied string
x.
}
\examples{

substrRight("hello world",3)
# returns "rld"

}
\seealso{
\code{\link{substrLeft}} for using n characters from the left side
of string,

\code{\link{substrMRight}} for removing n characters from the right side of
a string,

\code{\link{substrMLeft}} for removing n characters from the left side of a
string
}
\keyword{right}
\keyword{string}
