% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.infill.R
\name{timeseries.infill}
\alias{timeseries.infill}
\title{Infill discontinuous time series with blank values}
\usage{
timeseries.infill(ts)
}
\arguments{
\item{ts}{valid xts time series}
}
\value{
continuous xts time series
}
\description{
Takes xts dataset, finds minimum interval between time stamps and 
returns a new regular interval xts with same data content, but NA values inbetween known 
data values
}
\details{
only handles data with minimum time interval of 1 day; 1,2,3,4,6,8, or 12 hrs
}
\examples{
  dates <-seq(as.POSIXct("2012-05-01 00:00:00"), length=731, by="day")
  mydata<-xts(rnorm(731),order.by=dates)
  mydata<-mydata[wday(index(mydata))!=4] # remove wednesdays
  out<-timeseries.infill(mydata)

}
\author{
James R. Craig, University of Waterloo
}
\keyword{infill}
\keyword{timeseries}
