% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch_log.R
\name{try_catch_log}
\alias{try_catch_log}
\title{Custom tryCatch that catches erros and warnings in a try anc catch and sends them to sentry
code based on https://github.com/aryoda/tryCatchLog}
\usage{
try_catch_log(
  expr,
  ...,
  silent.warnings = TRUE,
  silent.messages = TRUE,
  tags = list(),
  extra = list(),
  messages.to.sentry = TRUE,
  warnings.to.sentry = TRUE,
  errors.to.sentry = TRUE
)
}
\arguments{
\item{expr}{function or expression that might generate the error}

\item{...}{Extra input that has to be passed to other functions}

\item{silent.warnings}{Boolean indicating wheter to print the warnings}

\item{silent.messages}{Boolean indicating wheter to print the messages}

\item{tags}{Named list of tags}

\item{extra}{Set extra context}

\item{messages.to.sentry}{Boolean indicating whether to send messages to Sentry}

\item{warnings.to.sentry}{Boolean indicating whether to send warnings to Sentry}

\item{errors.to.sentry}{Boolean indicating whether to send errors to Sentry}
}
\description{
\code{try_catch_log} Catches errors and warnings and sends them to sentry
}
