% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.read.R
\name{watershed.read}
\alias{watershed.read}
\title{Read in Raven WatershedStorage file}
\usage{
watershed.read(ff = NA)
}
\arguments{
\item{ff}{full file path to the WatershedStorage.csv file}
}
\value{
\item{watershed.storage}{data frame from the file with standardized
names} \item{units}{vector corresponding to units of each column}
}
\description{
watershed.read is used to read in the WatershedStorage.csv file produced by
the modelling Framework Raven.
}
\details{
This function expects a full file path to the WatershedStorage.csv file,
then reads in the file using read.csv. The main advantage of this functon is
renaming the columns to nicer names and extracting the units into something
much easier to read.

This function is also built to support the wshd.animate function, which uses
the object created here for creating an animation of the watershed storage
containers.

ff is the full file path of the WatershedStorage.csv file. If the file is
located in the current working directory, then simply the name of the file
is sufficient.
}
\examples{

# locate in RavenR Watershed storage file
ff <- system.file("extdata","run1_WatershedStorage.csv", package="RavenR")

# create full file path and read in file
mywshd <- watershed.read(ff)

# check data
head(mywshd$watershed.storage)

}
\seealso{
\code{\link{hyd.read}} for reading in the Hydrographs.csv file
\code{\link{watershedmeb.read}} for reading in the
WatershedMassEnergyBalance.csv file

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{read.csv}
\keyword{storage}
\keyword{watershed}
