% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.max.xts.R
\name{which.max.xts}
\alias{which.max.xts}
\title{Whixh Max (xts)}
\usage{
which.max.xts(x)
}
\arguments{
\item{x}{xts object to apply which.max to}
}
\value{
\item{x[which.max(x)]}{returned object}
}
\description{
which.max.xts applies the which.max function and returns an xts object
with the maximum value and associated date.
}
\details{
This function is intended to act as the which.max function, applicable
to xts objects and returning values in an xts format.
}
\examples{
data(hydrograph.data)
which.max.xts(hydrograph.data$hyd$Sub43_obs)

}
\seealso{
\code{\link{which.max}} base which.max function

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{max}
\keyword{which}
\keyword{xts}
