% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_file_write_utils.R
\name{write.RavenCommand}
\alias{write.RavenCommand}
\title{Writes common Raven command/parameter line to file (appends)}
\usage{
write.RavenCommand(command, value, filename, digits = NULL)
}
\arguments{
\item{command}{character, Raven command}

\item{value}{numeric or character, corresponding value}

\item{filename}{character, file name/path to write command/parameter line to, with extension}

\item{digits}{Number of digits to round value to (optional)}
}
\description{
Writes common Raven command/parameter line to file (appends)
}
\examples{
\dontrun{}
write.RavenCommand('Duration', 365, 'Hogwarts.rvi')
}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
