% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_file_write_utils.R
\name{write.RavenLabel}
\alias{write.RavenLabel}
\title{Writes common Raven labeled line to file (appends)}
\usage{
write.RavenLabel(label, filename)
}
\arguments{
\item{label}{character, (e.g. "SoilClasses")}

\item{filename}{character, file name/path to write label to, with extension}
}
\description{
Writes common Raven labeled line to file (appends)
}
\examples{
\dontrun{}
write.RavenLabel('SoilClasses', 'Hogwarts.rvp')
}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
