% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_file_write_utils.R
\name{write.RavenTable}
\alias{write.RavenTable}
\title{Writes a nicely formatted tables of Raven attributes/parameters}
\usage{
write.RavenTable(
  filename,
  attributes,
  units,
  df,
  justify = "right",
  sep = ", ",
  ...
)
}
\arguments{
\item{filename}{Name of the file, with extension, to append the table to}

\item{attributes}{array of strings containing attribute/parameter names}

\item{units}{array of strings with the corresponding units}

\item{df}{Dataframe of values corresponding to attributes/parameters}

\item{justify}{alignment of character columns (default 'right'). See \code{\link{format}}}

\item{sep}{character(s) used to seperate columns (default ', ')}

\item{...}{Extra arguments for \code{\link{write.fwf}}}
}
\description{
Writes a nicely formatted tables of Raven attributes/parameters
}
\examples{
\dontrun{}
soil_classes <- data.frame('Attributes' = c('DEFAULT','ALTERNATIVE'),
                           'SAND'      = c(0.4316, 0.3000),
                           'CLAY'      = c(0.1684, 0.4000),
                           'SILT'      = c(0.4000, 0.3000),
                           'ORGANIC'   = c(0.0000, 0.0000))
attributes <- c('\%SAND','\%CLAY','\%SILT','\%ORGANIC')
units <-  rep('none',4)
write.RavenTable('Hogwarts.rvp', attributes = attributes, units = units, df = soil_classes)

}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
