% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_file_write_utils.R
\name{writeRavenRFileInfo}
\alias{writeRavenRFileInfo}
\title{Write common Raven file header. All lines are Appended.}
\usage{
writeRavenRFileInfo(
  filename,
  filetype,
  author = NULL,
  creationDate = TRUE,
  textlen = 40
)
}
\arguments{
\item{filename}{Name of the file, with extension}

\item{filetype}{File extension, Encoding, Raven version (e.g. "rvp ASCII Raven 2.9.1")}

\item{author}{Name of file author (optional)}

\item{creationDate}{Bool of whether creation date should be added to header. (default TRUE)}

\item{textlen}{Length of lines (default: 40, used to right-align text)}
}
\description{
Write common Raven file header. All lines are Appended.
}
\examples{
\dontrun{}
writeRavenRFileInfo(filename = 'HogwartsBasin.rvp',
                    filetype = 'rvp ASCII Raven 2.9.1',
                    author   = 'Harry Potter')

}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
