% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyear.indices.aus.R
\name{wyear.indices.aus}
\alias{wyear.indices.aus}
\title{Water Year Indices (Australian)}
\usage{
wyear.indices.aus(sim)
}
\arguments{
\item{sim}{time series object to obtain indices for}
}
\value{
\item{wyear.ind}{array of indices corresponding to start/end of
water years}
}
\description{
wyear.indices.aus returns the indices of the provided time series for the
start/end of the water year (hardcoded as July 1st). Useful in evaluating
the time series in various diagnostics.
}
\details{
This function will return the indices for July 1st for each year in the time
series object. If the series ends on a June 30th date, it will return the
index for June 30th. If it starts or ends on a date that is not July 1st or
June 30th, that portion of the data will be disregarded (index not return
for that period).

The sim and obs should be of time series (xts) format.
}
\examples{

# read in sample forcings data
data("forcing.data")
fdata <- forcing.data$forcings

# get the indices of the start of the water year
wyear.indices.aus(fdata[,1])

}
\seealso{
\code{\link{wyear.indices}} for North American water year

\code{\link{flow.scatterplot}} for creating flow scatterplots

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Australia}
\keyword{Raven}
\keyword{indices}
\keyword{water}
\keyword{year}
