% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyear.indices.custom.R
\name{wyear.indices.custom}
\alias{wyear.indices.custom}
\title{Water Year Indices Custom}
\usage{
wyear.indices.custom(sim, mm, dd)
}
\arguments{
\item{sim}{time series object to obtain indices for}

\item{mm}{integer of month defined as start of the water year}

\item{dd}{integer of day defined as start of of the water year}
}
\value{
\item{wyear.ind}{array of indices corresponding to start/end of
water years}
}
\description{
wyear.indices.custom returns the indices of the provided time series for the
start/end of the water year, as defined by the user.
}
\details{
This function will return the indices for start of the water year, as
defined by the user using the mm and dd inputs. Note that unlike the
hardcoded functions, the day before the start of the water year will not be
searched as well.

The sim and obs should be of time series (xts) format.
}
\examples{

# read in sample forcings data
data("forcing.data")
fdata <- forcing.data$forcings

# get the indices of the start of the water year defined as April 1st
wyear.indices.custom(fdata[,1],4,1)

}
\seealso{
\code{\link{wyear.indices.aus}} for Australian water year

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{custom}
\keyword{indices}
\keyword{water}
\keyword{year}
