% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.wyearly.cols.R
\name{apply.wyearly.cols}
\alias{apply.wyearly.cols}
\title{Apply function for water year}
\usage{
apply.wyearly.cols(x, FUN, ...)
}
\arguments{
\item{x}{xts data frame to calculate FUN for}

\item{FUN}{the function to be applied}

\item{...}{optional arguments to FUN}
}
\description{
apply.wyearly.cols is a wrapper for apply.wyearly, and can be applied to
multiple columns of an xts data frame simultaneously.
}
\details{
apply.wyearly.cols calculates a function FUN for the periods defined by the
water year, for each column in x. This similar to other functions of the
form apply.<time period>, for example apply.daily, apply.monthly, etc. This
is a function especially helpful to hydrology data or results.

The function assumes an October 1st start to the water year. If the data
supplied is from, e.g. 2008-10-01 to 2009-09-30, the function will use the
September 30th date as the final day to apply the function for the water
year.

Note that currently the water year calculation is done from Oct 2 to Oct 1st
(or September 30th) inclusive. For example in the Oct 1 2007 to Oct 1 2008
water year, the end date is shown as "2008-10-01", and the function will be
evaluated for the period 2007-10-02 - 2008-10-01, inclusive.
}
\examples{

# read in forcing data (use sample data from package0
data(forcing.data)
myforcings <- forcing.data

# apply max function for a subset of forcings data; apply to multuple columns
apply.wyearly.cols(myforcings$forcings[,2:6],max,na.rm=T)

}
\seealso{
\code{\link{apply.wyearly}} for applying this function to a single
vector for most functions

See also \href{http://www.civil.uwaterloo.ca/jrcraig/software.html}{James R.
Craig's research page} for software downloads
}
\keyword{apply}
\keyword{water}
\keyword{year}
