% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_exception.R
\name{capture_exception}
\alias{capture_exception}
\title{Capture Exception}
\usage{
capture_exception(object, e, extra)
}
\arguments{
\item{object}{A Sentry client}

\item{e}{exception to catch}

\item{extra}{set extra context}
}
\description{
\code{capture_exception} Captures an exception and sends it to Sentry
}
