% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_exception.R
\name{capture_exception.sentry}
\alias{capture_exception.sentry}
\title{Capture Exception}
\usage{
\method{capture_exception}{sentry}(object, exception, extra = NULL)
}
\arguments{
\item{object}{A Sentry client}

\item{exception}{exception to catch}

\item{extra}{extra data to send}
}
\description{
\code{capture_exception.sentry} Captures an exception and sends it to Sentry
}
