% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_exception.R
\name{capture_exception.sentry}
\alias{capture_exception.sentry}
\title{Capture Exception}
\usage{
capture_exception.sentry(
  exception,
  extra = NULL,
  level = "error",
  tags = NULL,
  include_session_info = TRUE
)
}
\arguments{
\item{exception}{exception to catch}

\item{extra}{set extra context}

\item{level}{set level, warning or error for example}

\item{tags}{named list of tags}

\item{include_session_info}{whether to send platform and package list, takes up to 1s or more}

\item{object}{A Sentry client}
}
\description{
\code{capture_exception.sentry} Captures an exception and sends it to Sentry
}
