% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_text.R
\name{capture_text}
\alias{capture_text}
\title{Capture Text}
\usage{
capture_text(
  text = NULL,
  extra = NULL,
  level = "error",
  tags = NULL,
  include_session_info = TRUE,
  exception = NULL
)
}
\arguments{
\item{text}{text to send}

\item{extra}{set extra context}

\item{level}{set level, warning or error for example}

\item{tags}{named list of tags}

\item{include_session_info}{whether to send platform and package list, takes up to 1s or more}

\item{exception}{list containing `type` and `value` describing the exception}

\item{object}{A Sentry client}
}
\description{
\code{capture_text} Captures text and sends it to Sentry
}
