% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.read.R
\name{custom.read}
\alias{custom.read}
\title{Read Raven Custom Output files}
\usage{
custom.read(ff = NA, no.runname = F)
}
\arguments{
\item{ff}{full file path to the custom output file}

\item{no.runname}{boolean for whether a runName is supplied, important for
parsing the filename}
}
\value{
\item{custom.out}{data frame with the custom output data stored as xts
object}
}
\description{
custom.read is used to read any Raven custom output file
}
\details{
custom.read parses the filename and predicts the file format accordingly, so
it is important to use the unmodified file names for this function. The use
(or not) of a runname is accounted for.

The returned object is a time series object (xts format), which can be used
to easily plot the time series data. The otpions of the custom output are
included in the rav.obj attributes.
}
\examples{

# find sample rvh file for Nith subwatershed
ff <- system.file("extdata","run1_SNOW_Daily_Average_ByHRU.csv", package="RavenR")

# extract and plot custom data
mycustomdata <- custom.read(ff)
plot(mycustomdata[,5],main='Daily Average SNOW - HRU 5')

}
\seealso{
\code{\link{customoutput.plot}} for plotting custom output

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{custom}
\keyword{output}
\keyword{read}
