% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ost.read.R
\name{Ost.read}
\alias{Ost.read}
\title{Read Data from Ostrich output}
\usage{
Ost.read(ff = "OstModel0.txt")
}
\arguments{
\item{ff}{full file path to the OstModel0.txt file; default 'OstModel0.txt'}
}
\value{
\item{Ost.dd}{Ostrich data used in plotting}
}
\description{
Ost.read is used to read in Ostrich output data and return it as a data
object. This function is a wrapper for the read.table function.
}
\examples{

# locate sample OstModel0.txt file in RavenR package
ff <- system.file("extdata","OstModel0.txt", package="RavenR")

# read in file
mydata <- Ost.read(ff)

# basic plot of parameter par_g_3 through iterations
plot(mydata$par_g_3)

}
\seealso{
\code{\link{Ost.plot}} for handling Raven hydrograph files
# warning: example not run, sample example for associated files only
Download Ostrich at
\href{http://www.eng.buffalo.edu/~lsmatott/Ostrich/OstrichMain.html}{Shawn
Matott's web page}

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for other software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Ostrich}
\keyword{calibration}
