% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_customoutput_plot.R
\name{rvn_custom_output_plot}
\alias{rvn_custom_output_plot}
\title{Plot Raven Custom Output}
\usage{
rvn_custom_output_plot(cust, IDs = NULL, prd = NULL)
}
\arguments{
\item{cust}{custom output object from custom.read}

\item{IDs}{(optional) array of HRU IDs, subbasin IDs, HRU Group names/IDs to
include in plots}

\item{prd}{(optional) period to use in plotting}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
rvn_custom_output_plot is used to plot the custom output from Raven
}
\details{
rvn_custom_output_plot plots the custom output from a Raven model. The custom
output should be first read in using the custom.read function.
}
\examples{

# read in custom output from sample data
data("custom.data")

# plot custom data
mycustomdata <- custom.data
rvn_custom_output_plot(mycustomdata,IDs=seq(1,32),prd="2002-10-01/2003-09-01")

}
\seealso{
\code{\link{rvn_custom_output_plot}} for plotting custom output

See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\keyword{Raven}
\keyword{custom}
\keyword{output}
\keyword{plot}
