% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_gen_gridweights.R
\name{rvn_gen_gridweights}
\alias{rvn_gen_gridweights}
\title{Generate Raven grid weights from shapefile}
\usage{
rvn_gen_gridweights(
  HRUshpfile,
  Gridshpfile,
  ValidHRUIDs,
  HRUIDcol = "HRU_ID",
  gridIDcol = "cellID",
  outfile = "GridWeights.rvt"
)
}
\arguments{
\item{HRUshpfile}{polygon shapefile of HRUs with data column containing HRU IDs (.shp extension expected)}

\item{Gridshpfile}{polygon shapefile of grid cells with data column containing cell IDs (.shp extension expected)}

\item{ValidHRUIDs}{a vector of valid HRU IDs in the model}

\item{HRUIDcol}{the name of the HRUshpfile polygon which contains the HRU IDs}

\item{gridIDcol}{the name of the Gridshpfile polygon which contains the cell IDs}

\item{outfile}{optional name of output Raven gaugeweights file}
}
\value{
\item{TRUE}{returns TRUE if executed properly. Also output GaugeWeights.rvi file}
}
\description{
Generates a Raven grid weights file given an HRU shapefile and a grid shapefile.
}
\details{
Generate grid weights file GaugeWeights.rvi given an HRU shapefile with HRU ID column
HRUIDcol (default 'HRU_ID') and a grid shapefile with ID column gridIDcol (default: 'cellID')
weights are determined by the areal overlap of grid cell g and HRU k, i.e.,
 wt[k][g]=area[k][g]/area[k]
where wt[k][g] is the weight of cell g in HRU k, area[k] is the total area of HRU k,
and area[k][g] is the area of HRU k that is within cell g.

By definition, the grid domain has to completely cover the HRU domain such that the sum of
wt[k][g] for any k, over all g, is 1.0

not alot of QA/Qc included - can fail due to bad netCDF file or inappropriate UTM zone;
 uses rgdal and rgeos libraries. *Both shapefiles should be in same projection!* Unfortunately, the
 accuracy of the gIntersection() routine leaves a bit to be desired...
}
\examples{

\dontrun{

# load example rvh file
nith <- system.file("extdata",'Nith.rvh', package = "RavenR")
rvh <- rvn_rvh_read(nith)

# define HRU shapefiles (needs to be added to library)
HRUshpfile <- system.file("extdata","Nith_HRU.shp",package = "RavenR")

# write grid shapefile from netcdf file
nithnc <- system.file("extdata","Nith.nc",package = "RavenR")
GridCells <- "Nith_GridCells.shp"
UTMzone <- 17
GRDshpfile <- rvn_netcdf_to_gridshp(nithnc,UTMzone,GridCells)

# generate .rvi file of grid weights
validHRUIDs <- rvh$HRUtable$ID
tmp <- rvn_gen_gridweights(HRUshpfile, GRDshpfile, validHRUIDs,
gridIDcol = 'GridIDs', HRUIDcol = "HRU_ID", outfile = "Nith_GridWeights.rvi")

}

}
\seealso{
\code{\link{rvn_netcdf_to_gridshp}} for converting netcdf files to grid shapefile format
}
\author{
James R. Craig, University of Waterloo, 2019
}
