% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_gen_gridweights.R
\name{rvn_gen_gridweights}
\alias{rvn_gen_gridweights}
\title{Generates a Raven grid weights file given an HRU shapefile and a grid shapefile}
\usage{
rvn_gen_gridweights(
  HRUshpfile,
  Gridshpfile,
  ValidHRUIDs,
  HRUIDcol = "HRU_ID",
  gridIDcol = "cellID",
  outfile = "GridWeights.rvt"
)
}
\description{
Generate grid weights file GaugeWeights.rvi given an HRU shapefile with HRU ID column
HRUIDcol (default 'HRU_ID') and a grid shapefile with ID column gridIDcol (default: 'cellID')
weights are determined by the areal overlap of grid cell g and HRU k, i.e.,
 wt[k][g]=area[k][g]/area[k]
where wt[k][g] is the weight of cell g in HRU k, area[k] is the total area of HRU k,
and area[k][g] is the area of HRU k that is within cell g.
}
\details{
By definition, the grid domain has to completely cover the HRU domain such that the sum of
wt[k][g] for any k, over all g, is 1.0

 @param HRUshpfile polygon shapefile of HRUs with data column containing HRU IDs (.shp extension expected)
 @param Gridshpfile polygon shapefile of grid cells with data column containing cell IDs (.shp extension expected)
 @param ValidHRUIDs a vector of valid HRU IDs in the model
 @param HRUIDcol the name of the HRUshpfile polugon which contains the HRU IDs
 @param gridIDcol the name of the Gridshpfile polugon which contains the cell IDs
 @param outfile optional name of output Raven gaugeweights file


 @return \item{TRUE}{returns TRUE if executed properly. Also output GaugeWeights.rvi file}

 @details not alot of QA/Qc included - can fail due to bad netCDF file or inappropriate UTM zone;
 uses rgdal and rgeos libraries. *Both shapefiles should be in same projection!* Unfortunately, the
 accuracy of the gIntersection() routine leaves a bit to be desired...

 @author James R. Craig, University of Waterloo, 2019

 @seealso \code{\link{rvn_netcdf_to_gridshp}} for converting netcdf files to grid shapefile format

 @seealso \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R. Craig's research page} for software downloads,
 including the \href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\examples{

 \dontrun{
  # netcdf.to.gridshp(2017052200.nc,11,GridCells)
  rvn_gen_gridweights(ModelHRUs.shp,GridCells.shp,ValidHRUIDs,HRUIDcol="ID")

  # example workflow
  #--------------------------------------------
  setwd("C:/temp/RedDeer/")

  HRUshpfile<-"./maps/RedDeerHRUs.shp" # needs to have single column referring to HRU ID (here presumed 'HRU_ID')
  GRDshpfile<-"./maps/RDPS_GridRedDeer2.shp"
  rvh<-rvh.read("reddeer.rvh")

  validHRUIDs<-rvh$HRUtable$ID

  tmp<-gridweights.gen(HRUshpfile,GRDshpfile, validHRUIDs, gridIDcol="GridIDs",HRUIDcol="HRU_ID",outfile="RDPSGridWeights.rvt")

  #dsn=normalizePath(dirname(HRUshpfile))
  #lay=substr(basename(HRUshpfile),1,nchar(basename(HRUshpfile))-4)
  #HRUshp<-readOGR(dsn = dsn,layer = lay) # returns SpatialPolygonsDataFrame

 }


}
\keyword{conversion}
\keyword{grid}
\keyword{netcdf}
\keyword{shapefile}
