% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_gridweights_write.R
\name{rvn_gridweights_write}
\alias{rvn_gridweights_write}
\title{Write Raven GridWeights file}
\usage{
rvn_gridweights_write(gridweights = NULL, outfile = "GridWeights.rvt")
}
\arguments{
\item{gridweights}{list with the number of HRUs, number of grid cells, and gridweights data frame}

\item{outfile}{file to write gridweights to}
}
\value{
\item{hyd}{data frame from the file with standardized names}
}
\description{
rvn_gridweights_write is used to write a Raven grid weights file.
}
\details{
This function write the gridweights to the specified grid weights text file.
}
\examples{

# load example rvh file
nith <- system.file("extdata",'Nith.rvh', package = "RavenR")
rvh <- rvn_rvh_read(nith)

# adjust HRU shapefile to one per subbasin for demonstration
rvh$HRUtable <- rvh$HRUtable[c(1,6,15,25),]
rvh$HRUtable$Area <- rvh$SBtable$Area
rvh$HRUtable$ID <- rvh$HRUtable$SBID

# define HRU shapefile (use subbasin shapefile for example)
HRUshpfile <- system.file("extdata","Nith_shapefile_sample.shp",package = "RavenR")

# write grid shapefile from netcdf file
nithnc <- system.file("extdata/Nith_era5_sample.nc", package="RavenR")
Gridshpfile <- file.path(tempdir(), "Nith_gridcells.shp")
myshp <- rvn_netcdf_to_gridshp(ncfile=nithnc, projID=26917, outshp=Gridshpfile)

# generate .rvt file of grid weights
ValidHRUIDs <- rvh$HRUtable$ID
tfout <- file.path(tempdir(), "Nith_GridWeights.rvt")
gw <- rvn_gen_gridweights(HRUshpfile, Gridshpfile, ValidHRUIDs,
gridIDcol = 'GridIDs', HRUIDcol = "subID")

# write the gridweights (gw) object to file
rvn_gridweights_write(gw, outfile=file.path(tempdir(), "Nith_GridWeights.rvt"))

}
\seealso{
\code{\link{rvn_gen_gridweights}} for generating a grid weights file from an HRU shapefile
and netcdf file.
}
