% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_met_recordplot.R
\name{rvn_met_recordplot}
\alias{rvn_met_recordplot}
\title{EC Climate Gauge Record Overlap Visualization}
\usage{
rvn_met_recordplot(metadata = NULL, stndata = NULL, variables = NULL)
}
\arguments{
\item{metadata}{tibble of the station meta-data from weathercan::stations_search()}

\item{stndata}{tibble of the station data from weathercan::weather_dl()}

\item{variables}{if using weathercan::weather_dl(), column names for variables of interest (currently only accepts 1 per call)}
}
\description{
This function plots the length of Environment Canada climate station records,
accessed via the weathercan package, to identify periods in which multiple
station records overlap.
}
\details{
Accepts outputs from either the stations_search() or weather_dl() functions
from the weathercan package and extracts the start and end dates of the record
for plotting.
}
\examples{
# Plot record lengths using outputs from weathercan::stations_search()
library(weathercan)
metadata = stations_search(coords=c(50.109,-120.787),interval='day',dist=150) # EC stations 150 km of Merritt, BC
metadata = metadata[metadata$start>=2000,] # subset stations with recent data
metadata = metadata[1:3,] # take only the first 3 stations for brevity

rvn_met_recordplot(metadata=metadata)

# Plot record lengths using outputs from weathercan::weather_dl()
library(weathercan)
ids = c(47408,51598,44927) # station_id for 3 EC stations used above
stndata=weather_dl(station_ids=ids,interval='day')

rvn_met_recordplot(stndata=stndata,variables = "total_precip") # compare records for a specific variable

}
