% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_netcdf_to_gridshp.R
\name{rvn_netcdf_to_gridshp}
\alias{rvn_netcdf_to_gridshp}
\title{Generate grid overlay from netCDF file}
\usage{
rvn_netcdf_to_gridshp(ncfile, UTMzone, outshp = "rvn_output_shapefile")
}
\arguments{
\item{ncfile}{netCDF file with latitude and longitude variables}

\item{UTMzone}{UTM zone for exported shapefile (integer)}

\item{outshp}{name of output shapefile prefix (i.e., no .shp extension, DEFAULT 'rvn_output_shapefile')}
}
\value{
\item{TRUE}{returns TRUE if executed properly. Also generates shapefile in folder}
}
\description{
Takes the latitude-longitude cell coordinates from a netCDF file (assumed to be named 'lat' and 'long')
generates an estimate of the grid polygons associated with each netCDF cell and exports this to a shapefile (outshp)
}
\details{
Note that the function can fail due to bad netCDF file or inappropriate UTM zone.
}
\note{
Sample netcdf file in RavenR package obtained from [GWF Cuizinart](https://tuna.cs.uwaterloo.ca/), metadata for the ERA5 downloaded sample data may
 be found at \url{https://gwf-uwaterloo.github.io/wamedex/#13079_iso.xml}.
}
\examples{

ncfile <- system.file("extdata/Nith_era5_sample.nc", package="RavenR")
UTMzone <- 17
outshp <- 'Nith_output_shapefile'
rvn_netcdf_to_gridshp(ncfile, UTMzone, outshp)


}
\seealso{
\code{\link{rvn_gen_gridweights}} for generating a shapefile of gridweights
}
\author{
James R. Craig, University of Waterloo, 2019
}
