% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_num_days_month}
\alias{rvn_num_days_month}
\title{Number of Days in Month}
\usage{
rvn_num_days_month(date)
}
\arguments{
\item{date}{object in date format}
}
\value{
\item{int}{number of days in the month}
}
\description{
rvn_num_days_month is used to calculate the number of days in the month
}
\details{
This method includes leap years if they exist in the specified month.
}
\examples{

rvn_num_days_month(as.Date("2016-02-05"))
# 29

rvn_num_days_month(as.Date("2017-01-17"))
# 31

}
\seealso{
\code{\link{rvn_num_days}} for calculating the number of days between
two dates

See original code on post in Stack Overflow
\href{http://stackoverflow.com/questions/6243088/find-out-the-number-of-days-of-a-month-in-rfind}{
the number of days in a month}
}
