% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_run.R
\name{rvn_run}
\alias{rvn_run}
\title{Run Raven Executable}
\usage{
rvn_run(
  fileprefix = NULL,
  indir = getwd(),
  ravenexe = NULL,
  outdir = NULL,
  rvc = NULL,
  rvt = NULL,
  rvp = NULL,
  rvh = NULL,
  showoutput = FALSE
)
}
\arguments{
\item{fileprefix}{file prefix for main Raven input files.}

\item{indir}{string path for Raven input files}

\item{ravenexe}{file path to Raven executable}

\item{outdir}{string path for Raven output files (optional)}

\item{rvc}{file path to specific rvc file (optional)}

\item{rvt}{file path to specific rvt file (optional)}

\item{rvp}{file path to specific rvp file (optional)}

\item{rvh}{file path to specific rvh file (optional)}

\item{showoutput}{boolean whether to show output in console (passed to show.output.on.console within system) (default FALSE)}
}
\value{
Returns output code from the system command when running Raven
}
\description{
Invokes shell to execute a Raven model.
}
\details{
Uses the \code{\link{shell}} command to run the Raven.exe command.

If the \code{fileprefix} is not supplied, the function will detect the rvi file automatically and
use that in the run (if exactly one rvi file exists in the directory).

The \code{indir} path must point to the main Raven input files, if they are not in the working directory.

The \code{ravenexe} must point to the Raven.exe file; if not supplied it will look for it in the RavenR/extdata
path that is saved to when using \code{\link{rvn_download}}.
}
\examples{
\dontrun{
## NOT RUN (data download + runs Raven.exe)
url<-"http://raven.uwaterloo.ca/files/RavenTutorialFiles.zip"
temploc <- tempdir()
destfile<-paste(temploc,"/RavenTutorialFiles.zip",sep="")
download.file(url,destfile)
destdir<-paste(temploc,"/RavenTutorialFiles",sep="")
dir.create(destdir)
unzip(zipfile=destfile,exdir=destdir)
file.remove(destfile)

# Irondeqoiut example
rvn_run(indir=paste(destdir,"/Irond",sep=""),
        showoutput=TRUE)
}

}
\seealso{
\code{\link{rvn_download}}
}
