% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvh_blankSBdf.R
\name{rvn_rvh_blankSBdf}
\alias{rvn_rvh_blankSBdf}
\title{Generate Blank Raven SubBasin DataFrame}
\usage{
rvn_rvh_blankSBdf(nSubBasins = 1)
}
\arguments{
\item{nSubBasins}{Number of SubBasins in model, used to determine number of rows in table (default = 1)}
}
\value{
data.frame of blank SubBasin properties to be filled in by user
}
\description{
Generates a blank data frame for Raven subbasin properties.
}
\examples{
SBtable <- rvn_rvh_blankSBdf(nSubBasins = 3)
SBtable

}
\seealso{
\code{\link{rvn_rvh_blankHRUdf}} to generate blank HRU data frame
}
\author{
Leland Scantlebury
}
