% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_rvi_connections.R
\name{rvn_rvi_connections}
\alias{rvn_rvi_connections}
\title{Generate Hydrological process connections list}
\usage{
rvn_rvi_connections(HPTable, ProcConDataFile = "RavenProcessConnections.dat")
}
\arguments{
\item{HPTable}{a data frame of hydrologic processes generated from the rvi.read() routine}

\item{ProcConDataFile}{(optional) filename/location of RavenProcesConnections.dat file}
}
\value{
Returns a dataframe of all of the process connections Includes the following data columns:
process type, algorithm, 'from' compartment, 'to' compartment, and conditional.
}
\description{
This routine reads in a hydrologic process list from rvi.read() and generates
the list of hydrologic process connections. It relies on a valid and up-to-date
RavenProcessConnections.dat file
}
\examples{
 \dontrun{
  rvi<-rvi.read("example.rvi")
  conn<-rvn_rvi_connections(rvi$HydProcTable,ProcConDataFile="RavenProcessConnections.dat")
  rvi.HPNetworkPlot(conn,pdfout="network.pdf")
  }
}
\seealso{
rvn_rvi_read
See also the \href{http://raven.uwaterloo.ca/}{Raven page}
}
\author{
James R. Craig, University of Waterloo
}
\keyword{Hydrologic}
\keyword{Processes}
\keyword{Raven}
\keyword{connections}
\keyword{rvi}
