% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_utilities.R
\name{rvn_substrMRight}
\alias{rvn_substrMRight}
\title{substring minus characters from the Right}
\usage{
rvn_substrMRight(x, n)
}
\arguments{
\item{x}{a string to manipulate}

\item{n}{number of characters to remove from the right side of the string}
}
\description{
rvn_substrMRight returns a string x with n characters removed from the right
side of the string.
}
\examples{

rvn_substrMRight("hello world",3)
# returns "hello wo"

}
\seealso{
\code{\link{rvn_substrRight}} for using n characters from the right
side of string

\code{\link{rvn_substrLeft}} for using n characters from the left side of
string

\code{\link{rvn_substrMLeft}} for removing n characters from the left side of a
string
}
