% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMap.plot.R
\name{SBMap.plot}
\alias{SBMap.plot}
\title{Plot Continuous Data Using Subbasin Shapefile}
\usage{
SBMap.plot(shpfilename, subIDcol, plot.date, cust.data,
  leg.title = "Legend", leg.pos = "bottomleft",
  normalize.data = FALSE, colour1 = "green", colour2 = "blue",
  num.classes = 5, invalid.stop = TRUE, basins.label = "subID",
  plot.title = "", plot.invalid = F)
}
\arguments{
\item{shpfilename}{filename of shapefile containing HRU polygons, with one column inidicating Raven HRU ID}

\item{subIDcol}{string of subbasin ID column in shapefile}

\item{plot.date}{string of date to plot in custom.data}

\item{cust.data}{custom data set as read in by custom.read, for daily by_subbasin data}

\item{leg.title}{text for legend title}

\item{leg.pos}{position of legend}

\item{normalize.data}{whether to normalize data by all cust.data (TRUE) or just the data for the given date (FALSE)}

\item{colour1}{string indicating which colour (text or code) to use on lower bound of range}

\item{colour2}{string indicating which colour (text or code) to use on upper bound of range}

\item{num.classes}{number of classes to use in legend. Does not change the actual display colours}

\item{invalid.stop}{whether to stop if invalid basins are found (TRUE) or just continue with a warning (FALSE)}

\item{basins.label}{label to put on basins, one of c('None,'subID','value') to show nothing, subbasinIDs, or actual plotted values}

\item{plot.title}{title across top of plot}

\item{plot.invalid}{boolean indicating whether to plot invalid basins in grey (currently disabled)}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
Plots Raven custom output into a subbasin map
}
\details{
Does not currently support discrete data such as land use. Ability to include invalid basins in grey is currently disabled.
}
\examples{

# Raw sample data
shpfilename <- system.file("extdata","Nith_shapefile_sample.shp",package="RavenR")

# Custom Output data from Raven for Nith basin
cust.data <- custom.read(system.file("extdata","run1_PRECIP_Daily_Average_BySubbasin.csv",
                                  package="RavenR"))

subIDcol <- 'subID' # attriute in shapefile with subbasin IDs
plot.date <- "2003-03-30" # date for which to plot custom data

# function call
SBMap.plot(shpfilename,subIDcol,plot.date,cust.data)

}
\seealso{
\code{\link{subbasinNetwork.plot}} to create network plots

See also the \href{http://raven.uwaterloo.ca/}{Raven web site}
See also \href{http://www.civil.uwaterloo.ca/jrcraig/}{James R.
Craig's research page} for software downloads, including the
\href{http://www.civil.uwaterloo.ca/jrcraig/Raven/Main.html}{Raven page}
}
\author{
James R. Craig, University of Waterloo

Robert Chlumsky
}
\keyword{map}
\keyword{plot}
\keyword{subbasin}
