% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvn_file_write_utils.R
\name{write.RavenLabel}
\alias{write.RavenLabel}
\title{Writes common Raven labeled line to file, with optional value (appends)}
\usage{
write.RavenLabel(label, filename, value = NULL, digits = NULL)
}
\arguments{
\item{label}{character, (e.g. "SoilClasses")}

\item{filename}{character, file name/path to write to, with extension}

\item{value}{numeric or character, corresponding value written after label (optional)}

\item{digits}{Number of digits to round value to (optional)}
}
\description{
Writes common Raven labeled line to file, with optional value (appends)
}
\examples{
\dontrun{}
write.RavenLabel('Duration', 'Hogwarts.rvi', value=365)
}
\author{
Leland Scantlebury, \email{leland@scantle.com}
}
